Initial Revamp
DB files moved to DB folder.
Some arrays in the Data file moved to their own files.
