+++ src/Databases/ConduitDB.js
@@ -1,0 +1,1269 @@
+export const conduitDB = [
+  {
+    guid: 0,
+    icon: "inv_axe_02",
+    name: "Unknown",
+  },
+
+  /* -------------------------------------------------------------------------- */
+  /*                                    Druid                                   */
+  /* -------------------------------------------------------------------------- */
+
+  /* ---------------------------- Potency Conduits ---------------------------- */
+
+  {
+    guid: 340616,
+    abilityIcon: "spell_nature_crystalball.jpg",
+    name: {
+      en: "Flash of Clarity",
+      cn: "明晰闪愈",
+      de: "Augenblick der Klarsicht",
+      fr: "Instant de lucidité",
+      ru: "Вспышка ясности",
+    },
+    type: "Potency",
+    class: "Restoration Druid",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 340621,
+    abilityIcon: "item_herbd.jpg",
+    name: {
+      en: "Floral Recycling",
+      cn: "花开花谢",
+      de: "Pflanzliche Wiederherstellung",
+      fr: "Recyclage floral",
+      ru: "Переработка цветов",
+    },
+    type: "Potency",
+    class: "Restoration Druid",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 340550,
+    abilityIcon: "spell_nature_ravenform.jpg",
+    name: {
+      en: "Ready for Anything",
+      cn: "准备万全",
+      de: "Auf alles gefasst",
+      fr: "Prêt à tout",
+      ru: "Готовность ко всему",
+    },
+    type: "Potency",
+    class: "Restoration Druid",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 340549,
+    abilityIcon: "ability_druid_flourish.jpg",
+    name: {
+      en: "Unstoppable Growth",
+      cn: "狂野生长",
+      de: "Unaufhaltsames Wachstum",
+      fr: "Croissance irrésistible",
+      ru: "Беспрестанный рост",
+    },
+    type: "Potency",
+    class: "Restoration Druid",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 341446,
+    abilityIcon: "ability_ardenweald_druid.jpg",
+    name: {
+      en: "Conflux of Elements",
+      cn: "元素汇流",
+      de: "Konflux der Elemente",
+      fr: "Confluence des éléments",
+      ru: "Слияние стихий",
+    },
+    type: "Potency",
+    class: "Restoration Druid",
+    covenant: "Night Fae",
+    show: true,
+  },
+  {
+    guid: 341378,
+    abilityIcon: "ability_bastion_druid.jpg",
+    name: {
+      en: "Deep Allegiance",
+      cn: "诚挚忠心",
+      de: "Tiefe Loyalität",
+      fr: "Allégeance indéfectible",
+      ru: "Беспредельная верность",
+    },
+    type: "Potency",
+    class: "Restoration Druid",
+    covenant: "Kyrian",
+    show: true,
+  },
+  {
+    guid: 341383,
+    abilityIcon: "ability_revendreth_druid.jpg",
+    name: {
+      en: "Endless Thirst",
+      cn: "无穷饥渴",
+      de: "Endloser Durst",
+      fr: "Soif éternelle",
+      ru: "Неутолимая жажда",
+    },
+    type: "Potency",
+    class: "Restoration Druid",
+    covenant: "Venthyr",
+    show: true,
+  },
+  {
+    guid: 341447,
+    abilityIcon: "ability_maldraxxus_druid.jpg",
+    name: {
+      en: "Evolved Swarm",
+      cn: "进化蜂群",
+      de: "Evolvierter Schwarm",
+      fr: "Essaim évolué",
+      ru: "Эволюционировавший рой",
+    },
+    type: "Potency",
+    class: "Restoration Druid",
+    covenant: "Necrolord",
+    show: true,
+  },
+
+  /* --------------------------- Endurance Conduits --------------------------- */
+
+  {
+    guid: 340529,
+    abilityIcon: "spell_nature_stoneclawtotem.jpg",
+    name: {
+      en: "Tough as Bark",
+      cn: "韧如树皮",
+      de: "Borkenhart",
+      fr: "Dur comme l’écorce",
+      ru: "Древесная жесткость",
+    },
+    type: "Endurance",
+    class: "Restoration Druid",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 340553,
+    abilityIcon: "ability_druid_tigersroar.jpg",
+    name: {
+      en: "Well-Honed Instincts",
+      cn: "精确本能",
+      de: "Geschärfte Instinkte",
+      fr: "Instinct infaillible",
+      ru: "Отточенные инстинкты",
+    },
+    type: "Endurance",
+    class: "Restoration Druid",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 340540,
+    abilityIcon: "ability_druid_markofursol.jpg",
+    name: {
+      en: "Ursine Vigor",
+      cn: "巨熊活力",
+      de: "Kraft des Bären",
+      fr: "Vigueur ursine",
+      ru: "Медвежья энергия",
+    },
+    type: "Endurance",
+    class: "Restoration Druid",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 340543,
+    abilityIcon: "spell_nature_healingway.jpg",
+    name: {
+      en: "Innate Resolve",
+      cn: "巨熊固有决心活力",
+      de: "Angeborene Entschlossenheit",
+      fr: "Résolution innée",
+      ru: "Врожденная стойкость",
+    },
+    type: "Endurance",
+    class: "Restoration Druid",
+    covenant: "",
+    show: true,
+  },
+
+  /* ---------------------------- Finesse Conduits ---------------------------- */
+
+  {
+    guid: 341280,
+    abilityIcon: "spell_nature_reincarnation.jpg",
+    name: {
+      en: "Born Anew",
+      cn: "重获新生",
+      de: "Neu geboren",
+      fr: "Seconde naissance",
+      ru: "Рождение заново",
+    },
+    type: "Finesse",
+    class: "Restoration Druid",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 341450,
+    abilityIcon: "spell_druid_stampedingroar_cat.jpg",
+    name: {
+      en: "Front of the Pack",
+      cn: "兽群先锋",
+      de: "Spitze des Rudels",
+      fr: "Tête de la meute",
+      ru: "Во главе стаи",
+    },
+    type: "Finesse",
+    class: "Restoration Druid",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 341451,
+    abilityIcon: "ability_hunter_huntervswild.jpg",
+    name: {
+      en: "Born of the Wilds",
+      cn: "荒野之子",
+      de: "Kind der Wildnis",
+      fr: "Enfant sauvage",
+      ru: "Дитя дикой природы",
+    },
+    type: "Finesse",
+    class: "Restoration Druid",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 340545,
+    abilityIcon: "spell_druid_tirelesspursuit.jpg",
+    name: {
+      en: "Tireless Pursuit",
+      cn: "不息追猎",
+      de: "Unermüdliche Verfolgung",
+      fr: "Poursuite inlassable",
+      ru: "Неустанная погоня",
+    },
+    type: "Finesse",
+    class: "Restoration Druid",
+    covenant: "",
+    show: true,
+  },
+
+  /* -------------------------------------------------------------------------- */
+  /*                                    Monk                                    */
+  /* -------------------------------------------------------------------------- */
+
+  /* ---------------------------- Potency Conduits ---------------------------- */
+
+  {
+    guid: 336773,
+    abilityIcon: "inv_inscription_deck_jadeserpent.jpg",
+    name: {
+      en: "Jade Bond",
+      cn: "翡翠羁绊",
+      de: "Jadeband",
+      fr: "Lien de jade",
+      ru: "Связь c Нефритовой змеей",
+    },
+    type: "Potency",
+    class: "Mistweaver Monk",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 337241,
+    abilityIcon: "inv_misc_gem_pearl_06.jpg",
+    name: {
+      en: "Nourishing Chi",
+      cn: "滋养真气",
+      de: "Nährendes Chi",
+      fr: "Chi nourrissant",
+      ru: "Укрепляющая ци",
+    },
+    type: "Potency",
+    class: "Mistweaver Monk",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 337099,
+    abilityIcon: "spell_monk_revival.jpg",
+    name: {
+      en: "Rising Sun Revival",
+      cn: "旭日还魂",
+      de: "Belebung der aufgehenden Sonne",
+      fr: "Renaissance du soleil levant",
+      ru: "Восстановление восходящего солнца",
+    },
+    type: "Potency",
+    class: "Mistweaver Monk",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 336812,
+    abilityIcon: "spell_nature_abolishmagic.jpg",
+    name: {
+      en: "Resplendent Mist",
+      cn: "璀璨迷雾",
+      de: "Prächtiger Nebel",
+      fr: "Brume resplendissante",
+      ru: "Сверкающий туман",
+    },
+    type: "Potency",
+    class: "Mistweaver Monk",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 337286,
+    abilityIcon: "ability_bastion_warrior.jpg",
+    name: {
+      en: "Strike with Clarity",
+      cn: "洞察一击",
+      de: "Schlag voller Klarheit",
+      fr: "Frappe de clarté",
+      ru: "Удар просветленных",
+    },
+    type: "Potency",
+    class: "Mistweaver Monk",
+    covenant: "Kyrian",
+    show: true,
+  },
+  {
+    guid: 337301,
+    abilityIcon: "spell_animarevendreth_buff.jpg",
+    name: {
+      en: "Imbued Reflections",
+      cn: "注能映像",
+      de: "Verzauberte Reflexionen",
+      fr: "Reflets imprégnés",
+      ru: "Усиленные отражения",
+    },
+    type: "Potency",
+    class: "Mistweaver Monk",
+    covenant: "Venthyr",
+    show: true,
+  },
+  {
+    guid: 337295,
+    abilityIcon: "spell_animamaldraxxus_nova.jpg",
+    name: {
+      en: "Bone Marrow Hops",
+      cn: "骨髓酒花",
+      de: "Knochenmarkhopfen",
+      fr: "Houblon à moelle",
+      ru: "Хмель из костного мозга",
+    },
+    type: "Potency",
+    class: "Mistweaver Monk",
+    covenant: "Necrolords",
+    show: true,
+  },
+  {
+    guid: 337303,
+    abilityIcon: "spell_animaardenweald_beam.jpg",
+    name: {
+      en: "Way of the Fae",
+      cn: "法夜之道",
+      de: "Weg der Fae",
+      fr: "Voie des Faë",
+      ru: "Путь ночного народца",
+    },
+    type: "Potency",
+    class: "Mistweaver Monk",
+    covenant: "Night Fae",
+    show: true,
+  },
+
+  /* --------------------------- Endurance Conduits --------------------------- */
+
+  {
+    guid: 336853,
+    abilityIcon: "inv_misc_food_vendor_blackpepper.jpg",
+    name: {
+      en: "Fortifying Ingredients",
+      cn: "壮胆配料",
+      de: "Stärkende Zutaten",
+      fr: "Ingrédients fortifiants",
+      ru: "Укрепляющие ингредиенты",
+    },
+    type: "Endurance",
+    class: "Mistweaver Monk",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 336632,
+    abilityIcon: "inv_misc_volatileearth.jpg",
+    name: {
+      en: "Grounding Breath",
+      cn: "定息活血",
+      de: "Erdender Atem",
+      fr: "Souffle isolant",
+      ru: "Заземляющее дыхание",
+    },
+    type: "Endurance",
+    class: "Mistweaver Monk",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 336379,
+    abilityIcon: "ability_monk_chiswirl.jpg",
+    name: {
+      en: "Harm Denial",
+      cn: "接木成林",
+      de: "Schadensabwendung",
+      fr: "Refus du mal",
+      ru: "Защита от вреда",
+    },
+    type: "Endurance",
+    class: "Mistweaver Monk",
+    covenant: "",
+    show: true,
+  },
+
+  /* ---------------------------- Finesse Conduits ---------------------------- */
+
+  {
+    guid: 336890,
+    abilityIcon: "inv_elemental_mote_air01.jpg",
+    name: {
+      en: "Dizzying Tumble",
+      cn: "失神倾倒",
+      de: "Schwindelerregende Akrobatik",
+      fr: "Chute étourdissante",
+      ru: "Ошеломляющий кувырок",
+    },
+    type: "Finesse",
+    class: "Mistweaver Monk",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 336884,
+    abilityIcon: "sha_ability_rogue_bloodyeye.jpg",
+    name: {
+      en: "Lingering Numbness",
+      cn: "骨软筋酥",
+      de: "Verweilende Taubheit",
+      fr: "Engourdissement persistant",
+      ru: "Остаточное оцепенение",
+    },
+    type: "Finesse",
+    class: "Mistweaver Monk",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 337078,
+    abilityIcon: "ability_monk_ridethewind.jpg",
+    name: {
+      en: "Swift Transference",
+      cn: "迅捷双分",
+      de: "Flinker Transfer",
+      fr: "Transfert rapide",
+      ru: "Стремительное перенаправление",
+    },
+    type: "Finesse",
+    class: "Mistweaver Monk",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 337084,
+    abilityIcon: "ability_titankeeper_phasing.jpg",
+    name: {
+      en: "Tumbling Technique",
+      cn: "身法决",
+      de: "Akrobatisches Abrollen",
+      fr: "Art de la chute",
+      ru: "Техника кувырка",
+    },
+    type: "Finesse",
+    class: "Mistweaver Monk",
+    covenant: "",
+    show: true,
+  },
+
+  /* -------------------------------------------------------------------------- */
+  /*                                   Paladin                                  */
+  /* -------------------------------------------------------------------------- */
+
+  /* ---------------------------- Potency Conduits ---------------------------- */
+
+  {
+    guid: 339570,
+    abilityIcon: "ability_priest_spiritoftheredeemer.jpg",
+    name: {
+      en: "Enkindled Spirit",
+      cn: "光明之魂",
+      de: "Entflammter Geist",
+      fr: "Esprit embrasé",
+      ru: "Воодушевленный дух",
+    },
+    type: "Potency",
+    class: "Holy Paladin",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 339984,
+    abilityIcon: "item_holyspark.jpg",
+    name: {
+      en: "Focused Light",
+      cn: "聚焦圣光",
+      de: "Fokussiertes Licht",
+      fr: "Lumière concentrée",
+      ru: "Сфокусированный свет",
+    },
+    type: "Potency",
+    class: "Holy Paladin",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 339712,
+    abilityIcon: "ability_priest_voidshift.jpg",
+    name: {
+      en: "Resplendent Light",
+      cn: "壮丽圣光",
+      de: "Prächtiges Licht",
+      fr: "Lumière resplendissante",
+      ru: "Сияющий Свет",
+    },
+    type: "Potency",
+    class: "Holy Paladin",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 339987,
+    abilityIcon: "achievement_admiral_of_the_light.jpg",
+    name: {
+      en: "Untempered Dedication",
+      cn: "痴迷殉道",
+      de: "Ungezügelte Hingabe",
+      fr: "Dévouement intact",
+      ru: "Неумеренная преданность",
+    },
+    type: "Potency",
+    class: "Holy Paladin",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 340218,
+    abilityIcon: "spell_animabastion_orb.jpg",
+    name: {
+      en: "Ringing Clarity",
+      cn: "明澈之音",
+      de: "Läutende Klarheit",
+      fr: "Clarté retentissante",
+      ru: "Звенящая ясность",
+    },
+    type: "Potency",
+    class: "Holy Paladin",
+    covenant: "Kyrian",
+    show: true,
+  },
+  {
+    guid: 340212,
+    abilityIcon: "spell_animarevendreth_groundstate.jpg",
+    name: {
+      en: "Hallowed Discernment",
+      cn: "圣灰洞察",
+      de: "Geheiligte Einsicht",
+      fr: "Discernement sacré",
+      ru: "Священная проницательность",
+    },
+    type: "Potency",
+    class: "Holy Paladin",
+    covenant: "Venthyr",
+    show: true,
+  },
+  {
+    guid: 340192,
+    abilityIcon: "spell_animamaldraxxus_missile.jpg",
+    name: {
+      en: "Righteous Might",
+      cn: "正义之力",
+      de: "Rechtschaffene Macht",
+      fr: "Puissance pieuse",
+      ru: "Мощь праведников",
+    },
+    type: "Potency",
+    class: "Holy Paladin",
+    covenant: "Necrolord",
+    show: true,
+  },
+  {
+    guid: 340185,
+    abilityIcon: "spell_animaardenweald_buff.jpg",
+    name: {
+      en: "The Long Summer",
+      cn: "漫漫长夏",
+      de: "Der lange Sommer",
+      fr: "Été prolongé",
+      ru: "Долгое лето",
+    },
+    type: "Potency",
+    class: "Holy Paladin",
+    covenant: "Night Fae",
+    show: true,
+  },
+
+  /* --------------------------- Endurance Conduits --------------------------- */
+
+  {
+    guid: 338741,
+    abilityIcon: "ability_racial_holyresistance.jpg",
+    name: {
+      en: "Divine Call",
+      cn: "神圣召唤",
+      de: "Göttlicher Ruf",
+      fr: "Appel divin",
+      ru: "Божественный зов",
+    },
+    type: "Endurance",
+    class: "Holy Paladin",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 339114,
+    abilityIcon: "ability_priest_cascade.jpg",
+    name: {
+      en: "Golden Path",
+      cn: "辉金之道",
+      de: "Goldener Pfad",
+      fr: "Voie dorée",
+      ru: "Золотой путь",
+    },
+    type: "Endurance",
+    class: "Holy Paladin",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 338787,
+    abilityIcon: "ability_vehicle_sonicshockwave.jpg",
+    name: {
+      en: "Shielding Words",
+      cn: "盾卫圣令",
+      de: "Abschirmende Worte",
+      fr: "Mots protecteurs",
+      ru: "Спасительное торжество",
+    },
+    type: "Endurance",
+    class: "Holy Paladin",
+    covenant: "",
+    show: true,
+  },
+
+  /* ---------------------------- Finesse Conduits ---------------------------- */
+
+  {
+    guid: 339316,
+    abilityIcon: "achievement_dungeon_heroic_gloryoftheraider.jpg",
+    name: {
+      en: "Echoing Blessings",
+      cn: "回响祝福",
+      de: "Widerhallende Segen",
+      fr: "Bénédictions résonnantes",
+      ru: "Эхо благословений",
+    },
+    type: "Finesse",
+    class: "Holy Paladin",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 339268,
+    abilityIcon: "inv_lightforgedmatrixability_lightsjudgment.jpgg",
+    name: {
+      en: "Light's Barding",
+      cn: "圣光骑甲",
+      de: "Geschirr des Lichts",
+      fr: "Caparaçon de la Lumière",
+      ru: "Доспехи скакуна Света",
+    },
+    type: "Finesse",
+    class: "Holy Paladin",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 339124,
+    abilityIcon: "spell_holy_mindsooth.jpg",
+    name: {
+      en: "Pure Concentration",
+      cn: "纯粹专注",
+      de: "Reine Konzentration",
+      fr: "Pure concentration",
+      ru: "Чистое сосредоточение",
+    },
+    type: "Finesse",
+    class: "Holy Paladin",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 339292,
+    abilityIcon: "spell_holy_painsupression.jpg",
+    name: {
+      en: "Wrench Evil",
+      cn: "扭转邪恶",
+      de: "Böses ausrotten",
+      fr: "Renvoi du mal",
+      ru: "Искоренение зла",
+    },
+    type: "Finesse",
+    class: "Holy Paladin",
+    covenant: "",
+    show: true,
+  },
+
+  /* -------------------------------------------------------------------------- */
+  /*                                   Priest                                   */
+  /* -------------------------------------------------------------------------- */
+
+  /* ---------------------------- Potency Conduits ---------------------------- */
+
+  {
+    guid: 337790,
+    abilityIcon: "spell_holy_spiritualguidence.jpg", // This is a blizzard misspelling, not mine. Just by the way.
+    name: {
+      en: "Exaltation",
+      cn: "欣喜",
+      de: "Verherrlichung",
+      fr: "Exaltation",
+      ru: "Экзальтация",
+    },
+    type: "Potency",
+    class: "Discipline Priest",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 337786,
+    abilityIcon: "spell_holy_blessedrecovery.jpg",
+    name: {
+      en: "Pain Transformation",
+      cn: "痛苦转换",
+      de: "Schmerztransformation",
+      fr: "Transformation de la douleur",
+      ru: "Обращение боли",
+    },
+    type: "Potency",
+    class: "Discipline Priest",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 337778,
+    abilityIcon: "ability_priest_spiritoftheredeemer.jpg",
+    name: {
+      en: "Shining Radiance",
+      cn: "闪耀光辉",
+      de: "Glänzendes Strahlen",
+      fr: "Radiance éclatante",
+      ru: "Яркое сияние",
+    },
+    type: "Potency",
+    class: "Discipline Priest",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 337891,
+    abilityIcon: "spell_holy_purify.jpg",
+    name: {
+      en: "Swift Penitence",
+      cn: "极效苦修",
+      de: "Schnelle Buße",
+      fr: "Prompte pénitence",
+      ru: "Быстрое покаяние",
+    },
+    type: "Potency",
+    class: "Discipline Priest",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 337914,
+    abilityIcon: "achievement_bg_returnxflags_def_wsg.jpg",
+    name: {
+      en: "Focused Mending",
+      cn: "专注愈合",
+      de: "Fokussierte Besserung",
+      fr: "Guérison focalisée",
+      ru: "Направленное исцеление",
+    },
+    type: "Potency",
+    class: "Holy Priest",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 338345,
+    abilityIcon: "ability_priest_bindingprayers.jpg",
+    name: {
+      en: "Holy Oration",
+      cn: "圣言演说",
+      de: "Heilige Ansprache",
+      fr: "Oraison sacrée",
+      ru: "Священная молитва",
+    },
+    type: "Potency",
+    class: "Holy Priest",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 337811,
+    abilityIcon: "inv_icon_wing06a.jpg",
+    name: {
+      en: "Lasting Spirit",
+      cn: "延续之魂",
+      de: "Fortwährender Geist",
+      fr: "Esprit tenace",
+      ru: "Терпеливый дух",
+    },
+    type: "Potency",
+    class: "Holy Priest",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 337947,
+    abilityIcon: "spell_holy_holybolt.jpg",
+    name: {
+      en: "Resonant Words",
+      cn: "共鸣圣言",
+      de: "Resonierende Worte",
+      fr: "Mots résonnants",
+      ru: "Резонирующие слова",
+    },
+    type: "Potency",
+    class: "Holy Priest",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 337966,
+    abilityIcon: "spell_animabastion_wave.jpg",
+    name: {
+      en: "Courageous Ascension",
+      cn: "骁勇晋升",
+      de: "Mutiger Aufstieg",
+      fr: "Ascension courageuse",
+      ru: "Отважное перерождение",
+    },
+    type: "Potency",
+    class: "Priest",
+    covenant: "Kyrian",
+    show: true,
+  },
+  {
+    guid: 338315,
+    abilityIcon: "spell_animarevendreth_debuff.jpg",
+    name: {
+      en: "Shattered Perceptions",
+      cn: "破碎认知",
+      de: "Erschütterte Wahrnehmung",
+      fr: "Perceptions brisées",
+      ru: "Расколотое восприятие",
+    },
+    type: "Potency",
+    class: "Priest",
+    covenant: "Venthyr",
+    show: true,
+  },
+  {
+    guid: 337979,
+    abilityIcon: "spell_animamaldraxxus_debuff.jpg",
+    name: {
+      en: "Festering Transfusion",
+      cn: "溃烂灌注",
+      de: "Schwärende Transfusion",
+      fr: "Transfusion purulente",
+      ru: "Гнилостное переливание",
+    },
+    type: "Potency",
+    class: "Priest",
+    covenant: "Necrolord",
+    show: true,
+  },
+  {
+    guid: 338305,
+    abilityIcon: "spell_animaardenweald_buff.jpg",
+    name: {
+      en: "Fae Fermata",
+      cn: "法夜延音",
+      de: "Feenfermate",
+      fr: "Point d’orgue faë",
+      ru: "Арденвельдская фермата",
+    },
+    type: "Potency",
+    class: "Priest",
+    covenant: "Night Fae",
+    show: true,
+  },
+  {
+    guid: 337715,
+    abilityIcon: "ability_priest_clarityofwill.jpg",
+    name: {
+      en: "Charitable Soul",
+      cn: "慷慨之魂",
+      de: "Großzügige Seele",
+      fr: "Âme charitable",
+      ru: "Щедрая душа",
+    },
+    type: "Endurance",
+    class: "Priest",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 337748,
+    abilityIcon: "spell_holy_restoration.jpg",
+    name: {
+      en: "Light's Inspiration",
+      cn: "圣光鼓舞",
+      de: "Inspiration des Lichts",
+      fr: "Inspiration de la Lumière",
+      ru: "Вдохновение Света",
+    },
+    type: "Endurance",
+    class: "Priest",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 337662,
+    abilityIcon: "spell_nature_invisibilty.jpg",
+    name: {
+      en: "Translucent Image",
+      cn: "通透影像",
+      de: "Durchscheinendes Abbild",
+      fr: "Image translucide",
+      ru: "Прозрачный образ",
+    },
+    type: "Endurance",
+    class: "Priest",
+    covenant: "",
+    show: true,
+  },
+
+  /* ---------------------------- Finesse Conduits ---------------------------- */
+
+  {
+    guid: 337707,
+    abilityIcon: "ability_pvp_innerrenewal.jpg",
+    name: {
+      en: "Clear Mind",
+      cn: "清醒心智",
+      de: "Klarer Kopf",
+      fr: "Esprit clair",
+      ru: "Ясный разум",
+    },
+    type: "Finesse",
+    class: "Priest",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 337954,
+    abilityIcon: "sha_ability_rogue_bloodyeye_nightborne.jpg",
+    name: {
+      en: "Mental Recovery",
+      cn: "精神复建",
+      de: "Mentale Erholung",
+      fr: "Récupération mentale",
+      ru: "Ментальное восстановление",
+    },
+    type: "Finesse",
+    class: "Priest",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 337678,
+    abilityIcon: "ability_priest_savinggrace.jpg",
+    name: {
+      en: "Move with Grace",
+      cn: "优雅步态",
+      de: "Anmutige Bewegung",
+      fr: "Démarche distinguée",
+      ru: "Небесная грация",
+    },
+    type: "Finesse",
+    class: "Priest",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 337762,
+    abilityIcon: "ability_priest_wordsofmeaning.jpg",
+    name: {
+      en: "Power Unto Others",
+      cn: "赋能灌注",
+      de: "Macht den Anderen",
+      fr: "Transfert de puissance",
+      ru: "Сила единства",
+    },
+    type: "Finesse",
+    class: "Priest",
+    covenant: "",
+    show: true,
+  },
+
+  /* -------------------------------------------------------------------------- */
+  /*                                   Shaman                                   */
+  /* -------------------------------------------------------------------------- */
+
+  /* ---------------------------- Potency Conduits ---------------------------- */
+
+  {
+    guid: 338329,
+    abilityIcon: "spell_shaman_improvedearthshield.jpg",
+    name: {
+      en: "Embrace of Earth",
+      cn: "大地之拥",
+      de: "Umarmung der Erde",
+      fr: "Étreinte de la terre",
+      ru: "Объятия Земли",
+    },
+    type: "Potency",
+    class: "Restoration Shaman",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 338343,
+    abilityIcon: "shaman_pvp_ripplingwaters.jpg",
+    name: {
+      en: "Heavy Rainfall",
+      cn: "瓢泼大雨",
+      de: "Schwerer Regen",
+      fr: "Averse torrentielle",
+      ru: "Могучий ливень",
+    },
+    type: "Potency",
+    class: "Restoration Shaman",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 338346,
+    abilityIcon: "spell_shaman_blessingofeternals.jpg",
+    name: {
+      en: "Nature's Reach",
+      cn: "自然专注",
+      de: "Fokus der Natur",
+      fr: "Focalisation de la nature",
+      ru: "Средоточие природы",
+    },
+    type: "Potency",
+    class: "Restoration Shaman",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 338339,
+    abilityIcon: "spell_holy_serendipity.jpg",
+    name: {
+      en: "Swirling Currents",
+      cn: "旋涡之流",
+      de: "Wirbelnde Strömungen",
+      fr: "Courants tourbillonnants",
+      ru: "Кружащие потоки",
+    },
+    type: "Potency",
+    class: "Restoration Shaman",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 339182,
+    abilityIcon: "ability_monk_forcesphere.jpg",
+    name: {
+      en: "Elysian Dirge",
+      cn: "极乐挽歌",
+      de: "Elysisches Klagelied",
+      fr: "Requiem élyséen",
+      ru: "Элизийская панихида",
+    },
+    type: "Potency",
+    class: "Restoration Shaman",
+    covenant: "Kyrian",
+    show: true,
+  },
+  {
+    guid: 339185,
+    abilityIcon: "warlock_charredremains.jpg",
+    name: {
+      en: "Lavish Harvest",
+      cn: "华丽收割",
+      de: "Üppige Ernte",
+      fr: "Moisson généreuse",
+      ru: "Щедрая жатва",
+    },
+    type: "Potency",
+    class: "Restoration Shaman",
+    covenant: "Venthyr",
+    show: true,
+  },
+  {
+    guid: 339186,
+    abilityIcon: "spell_animamaldraxxus_wave.jpg",
+    name: {
+      en: "Tumbling Waves",
+      cn: "翻腾之潮",
+      de: "Stürzende Wellen",
+      fr: "Fracas des vagues",
+      ru: "Волна за волной",
+    },
+    type: "Potency",
+    class: "Restoration Shaman",
+    covenant: "Necrolord",
+    show: true,
+  },
+  {
+    guid: 339183,
+    abilityIcon: "spell_animaardenweald_debuff.jpg",
+    name: {
+      en: "Essential Extraction",
+      cn: "精粹提取",
+      de: "Essenzielle Extraktion",
+      fr: "Extraction essentielle",
+      ru: "Извлечение сути",
+    },
+    type: "Potency",
+    class: "Restoration Shaman",
+    covenant: "Night Fae",
+    show: true,
+  },
+
+  /* --------------------------- Endurance Conduits --------------------------- */
+
+  {
+    guid: 337964,
+    abilityIcon: "ability_hunter_mastertactitian.jpg",
+    name: {
+      en: "Astral Protection",
+      cn: "星界守护",
+      de: "Astralschutz",
+      fr: "Protection astrale",
+      ru: "Астральная защита",
+    },
+    type: "Endurance",
+    class: "Restoration Shaman",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 337974,
+    abilityIcon: "ability_shaman_fortifyingwaters.jpg",
+    name: {
+      en: "Refreshing Waters",
+      cn: "复苏水域",
+      de: "Erfrischende Wasser",
+      fr: "Eaux rafraîchissantes",
+      ru: "Освежающие воды",
+    },
+    type: "Endurance",
+    class: "Restoration Shaman",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 337981,
+    abilityIcon: "ability_accretion.jpg",
+    name: {
+      en: "Vital Accretion",
+      cn: "活力冲积",
+      de: "Vitalzuwachs",
+      fr: "Accrétion vitale",
+      ru: "Прирост жизненных сил",
+    },
+    type: "Endurance",
+    class: "Restoration Shaman",
+    covenant: "",
+    show: true,
+  },
+
+  /* ---------------------------- Finesse Conduits ---------------------------- */
+
+  {
+    guid: 338054,
+    abilityIcon: "spell_nature_removedisease.jpg",
+    name: {
+      en: "Crippling Hex",
+      cn: "致残妖术",
+      de: "Verkrüppelnde Verhexung",
+      fr: "Maléfice affaiblissant",
+      ru: "Увечащий сглаз",
+    },
+    type: "Finesse",
+    class: "Restoration Shaman",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 338048,
+    abilityIcon: "spell_arcane_prismaticcloak.jpg",
+    name: {
+      en: "Spiritual Resonance",
+      cn: "灵魂共鸣",
+      de: "Spirituelle Resonanz",
+      fr: "Résonance spirituelle",
+      ru: "Духовный резонанс",
+    },
+    type: "Finesse",
+    class: "Restoration Shaman",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 338033,
+    abilityIcon: "ability_hunter_longevity.jpg",
+    name: {
+      en: "Thunderous Paws",
+      cn: "雷霆之爪",
+      de: "Donnernde Pfoten",
+      fr: "Pattes tonitruantes",
+      ru: "Громоподобные лапы",
+    },
+    type: "Finesse",
+    class: "Restoration Shaman",
+    covenant: "",
+    show: true,
+  },
+  {
+    guid: 338042,
+    abilityIcon: "spell_shaman_dropall_03.jpg",
+    name: {
+      en: "Totemic Surge",
+      cn: "图腾潮涌",
+      de: "Kraft der Totems",
+      fr: "Afflux totémique",
+      ru: "Сила тотемов",
+    },
+    type: "Finesse",
+    class: "Restoration Shaman",
+    covenant: "",
+    show: true,
+  },
+];
