+++ src/Databases/ExternalsDB.js
@@ -1,0 +1,70 @@
+export const externalsDB = [
+  {
+    guid: 47788,
+    name: {
+      en: "Guardian Spirit",
+      de: "Schutzgeist",
+      fr: "Esprit gardien",
+      ru: "Оберегающий дух",
+      ch: "守护之魂",
+    },
+    wowhead: "https://www.wowhead.com/spell=47788/guardian-spirit",
+    icon: "spell_holy_guardianspirit",
+    cooldown: 180,
+  },
+
+  {
+    guid: 6940,
+    name: {
+      en: "Blessing of Sacrifice",
+      de: "Segen der Aufopferung",
+      fr: "Bénédiction de sacrifice",
+      ru: "Жертвенное благословение",
+      ch: "牺牲祝福",
+    },
+    wowhead: "https://www.wowhead.com/spell=6940/blessing-of-sacrifice",
+    icon: "spell_holy_sealofsacrifice",
+    cooldown: 120,
+  },
+  {
+    guid: 102342,
+    name: {
+      en: "Ironbark",
+      de: "Eisenborke",
+      fr: "Ecorcefer",
+      ru: "Железная кора",
+      ch: "铁木树皮",
+    },
+    wowhead: "https://www.wowhead.com/spell=102342/ironbark",
+    icon: "spell_druid_ironbark",
+    cooldown: 90,
+  },
+
+  {
+    guid: 10060,
+    name: {
+      en: "Power Infusion",
+      de: "Seele der Macht",
+      fr: "Infusion de puissance",
+      ru: "Придание сил",
+      ch: "能量灌注",
+    },
+    wowhead: "https://www.wowhead.com/spell=10060/power-infusion",
+    icon: "spell_holy_powerinfusion",
+    cooldown: 120,
+  },
+
+  {
+    guid: 116849,
+    name: {
+      en: "Life Cocoon",
+      de: "Lebenskokon",
+      fr: "Cocon de vie",
+      ru: "Исцеляющий кокон",
+      ch: "作茧缚命",
+    },
+    wowhead: "https://www.wowhead.com/spell=116849/life-cocoon",
+    icon: "ability_monk_chicocoon",
+    cooldown: 120,
+  },
+];
