+++ src/Databases/ServerDB.js
@@ -1,0 +1,952 @@
+export const serverDB = {
+  /* -------------------------------------------------------------------------- */
+  /*                                 US Servers                                 */
+  /* -------------------------------------------------------------------------- */
+
+  US: [
+    "Aegwynn",
+    "Aerie Peak",
+    "Agamaggan",
+    "Aggramar",
+    "Akama",
+    "Alexstrasza",
+    "Alleria",
+    "Altar of Storms",
+    "Alterac Mountains",
+    "Aman'Thul",
+    "Andorhal",
+    "Anetheron",
+    "Antonidas",
+    "Anub'arak",
+    "Anvilmar",
+    "Arathor",
+    "Archimonde",
+    "Area 52",
+    "Argent Dawn",
+    "Arthas",
+    "Arygos",
+    "Auchindoun",
+    "Azgalor",
+    "Azjol-Nerub",
+    "Azralon",
+    "Azshara",
+    "Azuremyst",
+    "Baelgun",
+    "Balnazzar",
+    "Barthilas",
+    "Black Dragonflight",
+    "Blackhand",
+    "Blackrock",
+    "Blackwater Raiders",
+    "Blackwing Lair",
+    "Blade's Edge",
+    "Bladefist",
+    "Bleeding Hollow",
+    "Blood Furnace",
+    "Bloodhoof",
+    "Bloodscalp",
+    "Bonechewer",
+    "Borean Tundra",
+    "Boulderfist",
+    "Bronzebeard",
+    "Burning Blade",
+    "Burning Legion",
+    "Caelestrasz",
+    "Cairne",
+    "Cenarion Circle",
+    "Cenarius",
+    "Cho'gall",
+    "Chromaggus",
+    "Coilfang",
+    "Crushridge",
+    "Daggerspine",
+    "Dalaran",
+    "Dalvengyr",
+    "Dark Iron",
+    "Darkspear",
+    "Darrowmere",
+    "Dath'Remar",
+    "Dawnbringer",
+    "Deathwing",
+    "Demon Soul",
+    "Dentarg",
+    "Destromath",
+    "Dethecus",
+    "Detheroc",
+    "Doomhammer",
+    "Draenor",
+    "Dragonblight",
+    "Dragonmaw",
+    "Drak'Tharon",
+    "Drak'thul",
+    "Draka",
+    "Drakkari",
+    "Dreadmaul",
+    "Drenden",
+    "Dunemaul",
+    "Durotan",
+    "Duskwood",
+    "Earthen Ring",
+    "Echo Isles",
+    "Eitrigg",
+    "Eldre'Thalas",
+    "Elune",
+    "Emerald Dream",
+    "Eonar",
+    "Eredar",
+    "Executus",
+    "Exodar",
+    "Farstriders",
+    "Feathermoon",
+    "Fenris",
+    "Firetree",
+    "Fizzcrank",
+    "Frostmane",
+    "Frostmourne",
+    "Frostwolf",
+    "Galakrond",
+    "Gallywix",
+    "Garithos",
+    "Garona",
+    "Garrosh",
+    "Ghostlands",
+    "Gilneas",
+    "Gnomeregan",
+    "Goldrinn",
+    "Gorefiend",
+    "Gorgonnash",
+    "Greymane",
+    "Grizzly Hills",
+    "Gul'dan",
+    "Gundrak",
+    "Gurubashi",
+    "Hakkar",
+    "Haomarush",
+    "Hellscream",
+    "Hydraxis",
+    "Hyjal",
+    "Icecrown",
+    "Illidan",
+    "Jaedenar",
+    "Jubei'Thos",
+    "Kael'thas",
+    "Kalecgos",
+    "Kargath",
+    "Kel'Thuzad",
+    "Khadgar",
+    "Khaz Modan",
+    "Khaz'goroth",
+    "Kil'jaeden",
+    "Kilrogg",
+    "Kirin Tor",
+    "Korgath",
+    "Korialstrasz",
+    "Kul Tiras",
+    "Laughing Skull",
+    "Lethon",
+    "Lightbringer",
+    "Lightning's Blade",
+    "Lightninghoof",
+    "Llane",
+    "Lothar",
+    "Madoran",
+    "Maelstrom",
+    "Magtheridon",
+    "Maiev",
+    "Mal'Ganis",
+    "Malfurion",
+    "Malorne",
+    "Malygos",
+    "Mannoroth",
+    "Medivh",
+    "Misha",
+    "Mok'Nathal",
+    "Moon Guard",
+    "Moonrunner",
+    "Mug'thol",
+    "Muradin",
+    "Nagrand",
+    "Nathrezim",
+    "Nazgrel",
+    "Nazjatar",
+    "Nemesis",
+    "Ner'zhul",
+    "Nesingwary",
+    "Nordrassil",
+    "Norgannon",
+    "Onyxia",
+    "Perenolde",
+    "Proudmoore",
+    "Quel'Thalas",
+    "Quel'dorei",
+    "Ragnaros",
+    "Ravencrest",
+    "Ravenholdt",
+    "Rexxar",
+    "Rivendare",
+    "Runetotem",
+    "Sargeras",
+    "Saurfang",
+    "Scarlet Crusade",
+    "Scilla",
+    "Sen'jin",
+    "Sentinels",
+    "Shadow Council",
+    "Shadowmoon",
+    "Shadowsong",
+    "Shandris",
+    "Shattered Halls",
+    "Shattered Hand",
+    "Shu'halo",
+    "Silver Hand",
+    "Silvermoon",
+    "Sisters of Elune",
+    "Skullcrusher",
+    "Skywall",
+    "Smolderthorn",
+    "Spinebreaker",
+    "Spirestone",
+    "Staghelm",
+    "Steamwheedle Cartel",
+    "Stonemaul",
+    "Stormrage",
+    "Stormreaver",
+    "Stormscale",
+    "Suramar",
+    "Tanaris",
+    "Terenas",
+    "Terokkar",
+    "Thaurissan",
+    "The Forgotten Coast",
+    "The Scryers",
+    "The Underbog",
+    "The Venture Co",
+    "Thorium Brotherhood",
+    "Thrall",
+    "Thunderhorn",
+    "Thunderlord",
+    "Tichondrius",
+    "Tol Barad",
+    "Tortheldrin",
+    "Trollbane",
+    "Turalyon",
+    "Twisting Nether",
+    "Uldaman",
+    "Uldum",
+    "Undermine",
+    "Ursin",
+    "Uther",
+    "Vashj",
+    "Vek'nilash",
+    "Velen",
+    "Warsong",
+    "Whisperwind",
+    "Wildhammer",
+    "Windrunner",
+    "Winterhoof",
+    "Wyrmrest Accord",
+    "Ysera",
+    "Ysondre",
+    "Zangarmarsh",
+    "Zul'jin",
+    "Zuluhed",
+  ],
+
+  /* -------------------------------------------------------------------------- */
+  /*                                 EU Servers                                 */
+  /* -------------------------------------------------------------------------- */
+
+  EU: [
+    "Aegwynn",
+    "Aerie Peak",
+    "Agamaggan",
+    "Aggra (Português)",
+    "Aggramar",
+    "Ahn'Qiraj",
+    "Al'Akir",
+    "Alexstrasza",
+    "Alleria",
+    "Alonsus",
+    "Aman'Thul",
+    "Ambossar",
+    "Anachronos",
+    "Anetheron",
+    "Antonidas",
+    "Anub'arak",
+    "Arak-arahm",
+    "Arathi",
+    "Arathor",
+    "Archimonde",
+    "Area 52",
+    "Argent Dawn",
+    "Arthas",
+    "Arygos",
+    "Ashenvale",
+    "Aszune",
+    "Auchindoun",
+    "Azjol-Nerub",
+    "Azshara",
+    "Azuregos",
+    "Azuremyst",
+    "Baelgun",
+    "Balnazzar",
+    "Blackhand",
+    "Blackmoore",
+    "Blackrock",
+    "Blackscar",
+    "Blade's Edge",
+    "Bladefist",
+    "Bloodfeather",
+    "Bloodhoof",
+    "Bloodscalp",
+    "Blutkessel",
+    "Booty Bay",
+    "Borean Tundra",
+    "Boulderfist",
+    "Bronze Dragonflight",
+    "Bronzebeard",
+    "Burning Blade",
+    "Burning Legion",
+    "Burning Steppes",
+    "C'Thun",
+    "Chamber of Aspects",
+    "Chants éternels",
+    "Cho'gall",
+    "Chromaggus",
+    "Colinas Pardas",
+    "Confrérie du Thorium",
+    "Conseil des Ombres",
+    "Crushridge",
+    "Culte de la Rive noire",
+    "Daggerspine",
+    "Dalaran",
+    "Dalvengyr",
+    "Darkmoon Faire",
+    "Darksorrow",
+    "Darkspear",
+    "Das Konsortium",
+    "Das Syndikat",
+    "Deathguard",
+    "Deathweaver",
+    "Deathwing",
+    "Deepholm",
+    "Defias Brotherhood",
+    "Dentarg",
+    "Der Mithrilorden",
+    "Der Rat von Dalaran",
+    "Der abyssische Rat",
+    "Destromath",
+    "Dethecus",
+    "Die Aldor",
+    "Die Arguswacht",
+    "Die Nachtwache",
+    "Die Silberne Hand",
+    "Die Todeskrallen",
+    "Die ewige Wacht",
+    "Doomhammer",
+    "Draenor",
+    "Dragonblight",
+    "Dragonmaw",
+    "Drak'thul",
+    "Drek'Thar",
+    "Dun Modr",
+    "Dun Morogh",
+    "Dunemaul",
+    "Durotan",
+    "Earthen Ring",
+    "Echsenkessel",
+    "Eitrigg",
+    "Eldre'Thalas",
+    "Elune",
+    "Emerald Dream",
+    "Emeriss",
+    "Eonar",
+    "Eredar",
+    "Eversong",
+    "Executus",
+    "Exodar",
+    "Festung der Stürme",
+    "Fordragon",
+    "Forscherliga",
+    "Frostmane",
+    "Frostmourne",
+    "Frostwhisper",
+    "Frostwolf",
+    "Galakrond",
+    "Garona",
+    "Garrosh",
+    "Genjuros",
+    "Ghostlands",
+    "Gilneas",
+    "Goldrinn",
+    "Gordunni",
+    "Gorgonnash",
+    "Greymane",
+    "Grim Batol",
+    "Grom",
+    "Gul'dan",
+    "Hakkar",
+    "Haomarush",
+    "Hellfire",
+    "Hellscream",
+    "Howling Fjord",
+    "Hyjal",
+    "Illidan",
+    "Jaedenar",
+    "Kael'thas",
+    "Karazhan",
+    "Kargath",
+    "Kazzak",
+    "Kel'Thuzad",
+    "Khadgar",
+    "Khaz Modan",
+    "Khaz'goroth",
+    "Kil'jaeden",
+    "Kilrogg",
+    "Kirin Tor",
+    "Kor'gall",
+    "Krag'jin",
+    "Krasus",
+    "Kul Tiras",
+    "Kult der Verdammten",
+    "La Croisade écarlate",
+    "Laughing Skull",
+    "Les Clairvoyants",
+    "Les Sentinelles",
+    "Lich King",
+    "Lightbringer",
+    "Lightning's Blade",
+    "Lordaeron",
+    "Los Errantes",
+    "Lothar",
+    "Madmortem",
+    "Magtheridon",
+    "Mal'Ganis",
+    "Malfurion",
+    "Malorne",
+    "Malygos",
+    "Mannoroth",
+    "Marécage de Zangar",
+    "Mazrigos",
+    "Medivh",
+    "Minahonda",
+    "Moonglade",
+    "Mug'thol",
+    "Nagrand",
+    "Nathrezim",
+    "Naxxramas",
+    "Nazjatar",
+    "Nefarian",
+    "Nemesis",
+    "Neptulon",
+    "Ner'zhul",
+    "Nera'thor",
+    "Nethersturm",
+    "Nordrassil",
+    "Norgannon",
+    "Nozdormu",
+    "Onyxia",
+    "Outland",
+    "Perenolde",
+    "Pozzo dell'Eternità",
+    "Proudmoore",
+    "Quel'Thalas",
+    "Ragnaros",
+    "Rajaxx",
+    "Rashgarroth",
+    "Ravencrest",
+    "Ravenholdt",
+    "Razuvious",
+    "Rexxar",
+    "Runetotem",
+    "Sanguino",
+    "Sargeras",
+    "Saurfang",
+    "Scarshield Legion",
+    "Sen'jin",
+    "Shadowsong",
+    "Shattered Halls",
+    "Shattered Hand",
+    "Shattrath",
+    "Shen'dralar",
+    "Silvermoon",
+    "Sinstralis",
+    "Skullcrusher",
+    "Soulflayer",
+    "Spinebreaker",
+    "Sporeggar",
+    "Steamwheedle Cartel",
+    "Stormrage",
+    "Stormreaver",
+    "Stormscale",
+    "Sunstrider",
+    "Suramar",
+    "Sylvanas",
+    "Taerar",
+    "Talnivarr",
+    "Tarren Mill",
+    "Teldrassil",
+    "Temple noir",
+    "Terenas",
+    "Terokkar",
+    "Terrordar",
+    "The Maelstrom",
+    "The Sha'tar",
+    "The Venture Co",
+    "Theradras",
+    "Thermaplugg",
+    "Thrall",
+    "Throk'Feroth",
+    "Thunderhorn",
+    "Tichondrius",
+    "Tirion",
+    "Todeswache",
+    "Trollbane",
+    "Turalyon",
+    "Twilight's Hammer",
+    "Twisting Nether",
+    "Tyrande",
+    "Uldaman",
+    "Ulduar",
+    "Uldum",
+    "Un'Goro",
+    "Varimathras",
+    "Vashj",
+    "Vek'lor",
+    "Vek'nilash",
+    "Vol'jin",
+    "Wildhammer",
+    "Wrathbringer",
+    "Xavius",
+    "Ysera",
+    "Ysondre",
+    "Zenedar",
+    "Zirkel des Cenarius",
+    "Zul'jin",
+    "Zuluhed",
+  ],
+
+  /* -------------------------------------------------------------------------- */
+  /*                               Korean Servers                               */
+  /* -------------------------------------------------------------------------- */
+
+  KR: [
+    "Alexstrasza",
+    "Azshara",
+    "Burning Legion",
+    "Cenarius",
+    "Dalaran",
+    "Deathwing",
+    "Durotan",
+    "Garona",
+    "Gul'dan",
+    "Hellscream",
+    "Hyjal",
+    "Malfurion",
+    "Norgannon",
+    "Rexxar",
+    "Stormrage",
+    "Wildhammer",
+    "Windrunner",
+    "Zul'jin",
+  ],
+
+  /* -------------------------------------------------------------------------- */
+  /*                               Taiwan Servers                               */
+  /* -------------------------------------------------------------------------- */
+
+  TW: [
+    "Arthas",
+    "Arygos",
+    "Bleeding Hollow",
+    "Chillwind Point",
+    "Crystalpine Stinger",
+    "Demon Fall Canyon",
+    "Dragonmaw",
+    "Frostmane",
+    "Hellscream",
+    "Icecrown",
+    "Light's Hope",
+    "Menethil",
+    "Nightsong",
+    "Order of the Cloud Serpent",
+    "Quel'dorei",
+    "Shadowmoon",
+    "Silverwing Hold",
+    "Skywall",
+    "Spirestone",
+    "Stormscale",
+    "Sundown Marsh",
+    "Whisperwind",
+    "World Tree",
+    "Wrathbringer",
+    "Zealot Blade",
+  ],
+
+  /* -------------------------------------------------------------------------- */
+  /*                               Chinese Servers                              */
+  /* -------------------------------------------------------------------------- */
+
+  CN: [
+    "万色星辰",
+    "世界之树",
+    "丹莫德",
+    "主宰之剑",
+    "丽丽（四川",
+    "亚雷戈斯",
+    "亡语者",
+    "伊兰尼库斯",
+    "伊利丹",
+    "伊森利恩",
+    "伊森德雷",
+    "伊瑟拉",
+    "伊莫塔尔",
+    "伊萨里奥斯",
+    "元素之力",
+    "克尔苏加德",
+    "克洛玛古斯",
+    "克苏恩",
+    "兰娜瑟尔",
+    "军团要塞",
+    "冬寒",
+    "冬拥湖",
+    "冬泉谷",
+    "冰川之拳",
+    "冰霜之刃",
+    "冰风岗",
+    "凤凰之神",
+    "凯尔萨斯",
+    "凯恩血蹄",
+    "刀塔",
+    "利刃之拳",
+    "刺骨利刃",
+    "加兹鲁维",
+    "加基森",
+    "加尔",
+    "加里索斯",
+    "勇士岛",
+    "千针石林",
+    "卡德加",
+    "卡德罗斯",
+    "卡扎克",
+    "卡拉赞",
+    "卡珊德拉",
+    "厄祖玛特",
+    "双子峰",
+    "古加尔",
+    "古尔丹",
+    "古拉巴什",
+    "古达克",
+    "哈兰",
+    "哈卡",
+    "嚎风峡湾",
+    "回音山",
+    "国王之谷",
+    "图拉扬",
+    "圣火神殿",
+    "地狱之石",
+    "地狱咆哮",
+    "埃克索图斯",
+    "埃加洛尔",
+    "埃基尔松",
+    "埃德萨拉",
+    "埃苏雷格",
+    "埃雷达尔",
+    "埃霍恩",
+    "基尔加丹",
+    "基尔罗格",
+    "塔伦米尔",
+    "塔纳利斯",
+    "塞拉摩",
+    "塞拉赞恩",
+    "塞泰克",
+    "塞纳留斯",
+    "壁炉谷",
+    "夏维安",
+    "外域",
+    "大地之怒",
+    "大漩涡",
+    "天空之墙",
+    "天谴之门",
+    "太阳之井",
+    "夺灵者",
+    "奈法利安",
+    "奈萨里奥",
+    "奎尔丹纳斯",
+    "奎尔萨拉斯",
+    "奥妮克希亚",
+    "奥尔加隆",
+    "奥拉基尔",
+    "奥斯里安",
+    "奥杜尔",
+    "奥特兰克",
+    "奥蕾莉亚",
+    "奥达曼",
+    "奥金顿",
+    "守护之剑",
+    "安东尼达斯",
+    "安其拉",
+    "安加萨",
+    "安威玛尔",
+    "安戈洛",
+    "安格博达",
+    "安纳塞隆",
+    "安苏",
+    "密林游侠",
+    "寒冰皇冠",
+    "尘风峡谷",
+    "屠魔山谷",
+    "山丘之王",
+    "巨龙之吼",
+    "巫妖之王",
+    "巴尔古恩",
+    "巴瑟拉斯",
+    "巴纳扎尔",
+    "布兰卡德",
+    "布莱克摩",
+    "布莱恩",
+    "布鲁塔卢斯",
+    "希尔瓦娜斯",
+    "希雷诺斯",
+    "幽暗沼泽",
+    "库尔提拉斯",
+    "库德兰",
+    "弗塞雷迦",
+    "影之哀伤",
+    "影牙要塞",
+    "德拉诺",
+    "恐怖图腾",
+    "恶魔之翼",
+    "恶魔之魂",
+    "戈古纳斯",
+    "戈提克",
+    "战歌",
+    "扎拉赞恩",
+    "托塞德林",
+    "托尔巴拉德",
+    "拉文凯斯",
+    "拉文霍德",
+    "拉格纳罗斯",
+    "拉贾克斯",
+    "提尔之手",
+    "提瑞斯法",
+    "摩摩尔",
+    "斩魔者",
+    "斯克提斯",
+    "斯坦索姆",
+    "无尽之海",
+    "无底海渊",
+    "日落沼泽",
+    "时光之穴",
+    "普瑞斯托",
+    "普罗德摩",
+    "晴日峰（江苏）",
+    "暗影之月",
+    "暗影裂口",
+    "暗影议会",
+    "暗影迷宫",
+    "暮色森林",
+    "暴风祭坛",
+    "月光林地",
+    "月神殿",
+    "末日祷告祭坛",
+    "末日行者",
+    "朵丹尼尔",
+    "杜隆坦",
+    "格瑞姆巴托",
+    "格雷迈恩",
+    "格鲁尔",
+    "桑德兰",
+    "梅尔加尼",
+    "梦境之树",
+    "森金",
+    "死亡之翼",
+    "死亡熔炉",
+    "毁灭之锤",
+    "永夜港",
+    "永恒之井",
+    "沃金",
+    "沙怒",
+    "法拉希姆",
+    "泰兰德",
+    "泰拉尔",
+    "洛丹伦",
+    "洛肯",
+    "洛萨",
+    "海克泰尔",
+    "海加尔",
+    "海达希亚",
+    "深渊之喉",
+    "深渊之巢",
+    "激流之傲",
+    "激流堡",
+    "火喉",
+    "火烟之谷",
+    "火焰之树",
+    "火羽山",
+    "灰谷",
+    "烈焰峰",
+    "烈焰荆棘",
+    "熊猫酒仙",
+    "熔火之心",
+    "熵魔",
+    "燃烧之刃",
+    "燃烧军团",
+    "燃烧平原",
+    "爱斯特纳",
+    "狂热之刃",
+    "狂风峭壁",
+    "玛多兰",
+    "玛法里奥",
+    "玛洛加尔",
+    "玛瑟里顿",
+    "玛维·影歌",
+    "玛诺洛斯",
+    "玛里苟斯",
+    "瑞文戴尔",
+    "瑟莱德丝",
+    "瓦丝琪",
+    "瓦拉斯塔兹",
+    "瓦拉纳",
+    "瓦里安",
+    "瓦里玛萨斯",
+    "甜水绿洲",
+    "生态船",
+    "白银之手",
+    "白骨荒野",
+    "盖斯",
+    "石爪峰",
+    "石锤",
+    "破碎岭",
+    "祖尔金",
+    "祖达克",
+    "祖阿曼",
+    "神圣之歌",
+    "穆戈尔",
+    "符文图腾",
+    "米奈希尔",
+    "索拉丁",
+    "索瑞森",
+    "红云台地",
+    "红龙军团",
+    "红龙女王",
+    "纳克萨玛斯",
+    "纳沙塔尔",
+    "织亡者",
+    "罗宁",
+    "罗曼斯",
+    "羽月",
+    "翡翠梦境",
+    "耐奥祖",
+    "耐普图隆",
+    "耳语海岸",
+    "能源舰",
+    "自由之风",
+    "艾森娜",
+    "艾欧娜尔",
+    "艾维娜",
+    "艾苏恩",
+    "艾莫莉丝",
+    "艾萨拉",
+    "艾露恩",
+    "芬里斯",
+    "苏塔恩",
+    "苏拉玛",
+    "范克里夫",
+    "范达尔鹿盔",
+    "荆棘谷",
+    "莱索恩",
+    "菲拉斯",
+    "菲米丝",
+    "萨尔",
+    "萨格拉斯",
+    "萨洛拉丝",
+    "萨菲隆",
+    "蓝龙军团",
+    "藏宝海湾",
+    "蜘蛛王国",
+    "血吼",
+    "血牙魔王",
+    "血环",
+    "血羽",
+    "血色十字军",
+    "血顶",
+    "试炼之环",
+    "诺兹多姆",
+    "诺森德",
+    "诺莫瑞根",
+    "贫瘠之地",
+    "踏梦者",
+    "轻风之语",
+    "辛达苟萨",
+    "达克萨隆",
+    "达基萨斯",
+    "达尔坎",
+    "达文格尔",
+    "达斯雷玛",
+    "达纳斯",
+    "达隆米尔",
+    "迅捷微风",
+    "远古海滩",
+    "迦拉克隆",
+    "迦玛兰",
+    "迦罗娜",
+    "迦顿",
+    "迪托马斯",
+    "迪瑟洛克",
+    "逐日者",
+    "通灵学院",
+    "遗忘海岸",
+    "金度",
+    "金色平原",
+    "铜龙军团",
+    "银月",
+    "银松森林",
+    "闪电之刃",
+    "阿克蒙德",
+    "阿努巴拉克",
+    "阿卡玛",
+    "阿古斯",
+    "阿尔萨斯",
+    "阿扎达斯",
+    "阿拉希",
+    "阿拉索",
+    "阿斯塔洛",
+    "阿曼尼",
+    "阿格拉玛",
+    "阿比迪斯",
+    "阿纳克洛斯",
+    "阿迦玛甘",
+    "雏龙之翼",
+    "雷克萨",
+    "雷斧堡垒",
+    "雷霆之怒",
+    "雷霆之王",
+    "雷霆号角",
+    "霍格",
+    "霜之哀伤",
+    "霜狼",
+    "风暴之怒",
+    "风暴之眼",
+    "风暴之鳞",
+    "风暴峭壁",
+    "风行者",
+    "鬼雾峰",
+    "鲜血熔炉",
+    "鹰巢山",
+    "麦姆",
+    "麦迪文",
+    "黄金之路",
+    "黑手军团",
+    "黑暗之矛",
+    "黑暗之门",
+    "黑暗虚空",
+    "黑暗魅影",
+    "黑石尖塔",
+    "黑翼之巢",
+    "黑铁",
+    "黑锋哨站",
+    "黑龙军团",
+    "龙骨平原",
+  ],
+};
