+++ src/Databases/SpellDB.js
@@ -1,0 +1,44 @@
+export const spellData = [
+  /* -------------------------------------------------------------------------- */
+  /*                                 Monk Spells                                */
+  /* -------------------------------------------------------------------------- */
+  { id: 124682, name: "Enveloping Mist", icon: "spell_monk_envelopingmist" },
+  { id: 116670, name: "Vivify", icon: "ability_monk_vivify" },
+  { id: 115310, name: "Revival", icon: "spell_monk_revival" },
+  { id: 191894, name: "Mastery: Gust of Mists", icon: "ability_monk_souldance" },
+  { id: 343655, name: "Enveloping Breath", icon: "ability_monk_dragonkick" },
+  { id: 191837, name: "Essence Font", icon: "ability_monk_essencefont" },
+  { id: 115151, name: "Renewing Mist", icon: "ability_monk_renewingmists" },
+  { id: 274909, name: "Rising Mist", icon: "ability_monk_effuse" },
+  { id: 116849, name: "Life Cocoon", icon: "ability_monk_chicocoon" },
+  { id: 115175, name: "Soothing Mist", icon: "ability_monk_soothingmists" },
+  { id: 310454, name: "Weapons of Order", icon: "ability_bastion_monk" },
+
+  /* -------------------------------------------------------------------------- */
+  /*                                Druid Spells                                */
+  /* -------------------------------------------------------------------------- */
+  { id: 774, name: "Rejuvenation", icon: "spell_nature_rejuvenation" },
+  { id: 48438, name: "Wild Growth", icon: "ability_druid_flourish" },
+  { id: 8936, name: "Regrowth", icon: "spell_nature_resistnature" },
+  { id: 33763, name: "Lifebloom", icon: "inv_misc_herb_felblossom" },
+
+  /* -------------------------------------------------------------------------- */
+  /*                                Shaman Spells                               */
+  /* -------------------------------------------------------------------------- */
+  { id: 61295, name: "Riptide", icon: "spell_nature_riptide" },
+  { id: 157153, name: "Cloudburst Totem", icon: "ability_shaman_condensationtotem" },
+  { id: 73920, name: "Healing Rain", icon: "spell_nature_giftofthewaterspirit" },
+  { id: 108280, name: "Healing Tide Totem", icon: "ability_shaman_healingtide" },
+  { id: 1064, name: "Chain Heal", icon: "spell_nature_healingwavegreater" },
+  { id: 16191, name: "Mana Tide Totem", icon: "spell_frost_summonwaterelemental" },
+  { id: 8004, name: "Healing Surge", icon: "spell_nature_healingway" },
+
+  /* -------------------------------------------------------------------------- */
+  /*                                Holy Paladin                                */
+  /* -------------------------------------------------------------------------- */
+  { id: 225311, name: "Light of Dawn", icon: "spell_paladin_lightofdawn" },
+  { id: 85673, name: "Word of Glory", icon: "inv_helmet_96" },
+  { id: 25914, name: "Holy Shock", icon: "spell_holy_searinglight" },
+  { id: 82326, name: "Holy Light", icon: "spell_holy_surgeoflight" },
+  { id: 337824, name: "Shock Barrier", icon: "ability_paladin_blessedmending" },
+];
