+++ src/Databases/TalentDB.js
@@ -1,0 +1,157 @@
+export default talentDB = [
+  /* -------------------------------------------------------------------------- */
+  /*                               Paladin Talents                              */
+  /* -------------------------------------------------------------------------- */
+
+  { name: "Crusader's Might", guid: 196926 },
+  { name: "Bestow Faith", guid: 223306 },
+  { name: "Light's Hammer", guid: 114158 },
+  { name: "Saved by the Light", guid: 157047 },
+  { name: "Judgment of Light", guid: 183778 },
+  { name: "Holy Prism", guid: 114165 },
+  { name: "Fist of Justice", guid: 234299 },
+  { name: "Repentance", guid: 20066 },
+  { name: "Blinding Light", guid: 115750 },
+  { name: "Unbreakable Spirit", guid: 114154 },
+  { name: "Cavalier", guid: 230332 },
+  { name: "Rule of Law", guid: 214202 },
+  { name: "Divine Purpose", guid: 223817 },
+  { name: "Holy Avenger", guid: 105809 },
+  { name: "Seraphim", guid: 152262 },
+  { name: "Sanctified Wrath", guid: 53376 },
+  { name: "Avenging Crusader", guid: 216331 },
+  { name: "Awakening", guid: 248033 },
+  { name: "Glimmer of Light", guid: 325966 },
+  { name: "Beacon of Faith", guid: 156910 },
+  { name: "Beacon of Virtue", guid: 200025 },
+
+  /* -------------------------------------------------------------------------- */
+  /*                               Shaman Talents                               */
+  /* -------------------------------------------------------------------------- */
+
+  { name: "Torrent", guid: 200072 },
+  { name: "Undulation", guid: 200071 },
+  { name: "Unleash Life", guid: 73685 },
+  { name: "Echo of the Elements", guid: 108283 },
+  { name: "Deluge", guid: 200076 },
+  { name: "Surge of Earth", guid: 320746 },
+  { name: "Spirit Wolf", guid: 260878 },
+  { name: "Earthgrab Totem", guid: 51485 },
+  { name: "Static Charge", guid: 265046 },
+  { name: "Ancestral Vigor", guid: 207401 },
+  { name: "Earthen Wall Totem", guid: 198838 },
+  { name: "Ancestral Protection Totem", guid: 207399 },
+  { name: "Nature's Guardian", guid: 30884 },
+  { name: "Graceful Spirit", guid: 192088 },
+  { name: "Wind Rush Totem", guid: 192077 },
+  { name: "Flash Flood", guid: 280614 },
+  { name: "Downpour", guid: 207778 },
+  { name: "Cloudburst Totem", guid: 157153 },
+  { name: "High Tide", guid: 157154 },
+  { name: "Wellspring", guid: 197995 },
+  { name: "Ascendance", guid: 114052 },
+
+  /* -------------------------------------------------------------------------- */
+  /*                                Druid Talents                               */
+  /* -------------------------------------------------------------------------- */
+
+  { name: "Abundance", guid: 207383 },
+  { name: "Nourish", guid: 50464 },
+  { name: "Cenarion Ward", guid: 102351 },
+  { name: "Tiger Dash", guid: 252216 },
+  { name: "Renewal", guid: 108238 },
+  { name: "Wild Charge", guid: 132302 },
+  { name: "Balance Affinity", guid: 197632 },
+  { name: "Feral Affinity", guid: 197490 },
+  { name: "Guardian Affinity", guid: 197491 },
+  { name: "Mighty Bash", guid: 5211 },
+  { name: "Mass Entanglement", guid: 102359 },
+  { name: "Heart of the Wild", guid: 319454 },
+  { name: "Soul of the Forest", guid: 158478 },
+  { name: "Cultivation", guid: 200390 },
+  { name: "Incarnation: Tree of Life", guid: 33891 },
+  { name: "Inner Peace", guid: 197073 },
+  { name: "Spring Blossoms", guid: 207385 },
+  { name: "Overgrowth", guid: 203651 },
+  { name: "Photosynthesis", guid: 274902 },
+  { name: "Germination", guid: 155675 },
+  { name: "Flourish", guid: 197721 },
+
+  /* -------------------------------------------------------------------------- */
+  /*                                Monk Talents                                */
+  /* -------------------------------------------------------------------------- */
+
+  { name: "Mist Wrap", guid: 197900 },
+  { name: "Chi Wave", guid: 115098 },
+  { name: "Chi Burst", guid: 123986 },
+  { name: "Celerity", guid: 115173 },
+  { name: "Chi Torpedo", guid: 115008 },
+  { name: "Tiger's Lust", guid: 116841 },
+  { name: "Lifecycles", guid: 197915 },
+  { name: "Spirit of the Crane", guid: 210802 },
+  { name: "Mana Tea", guid: 197908 },
+  { name: "Tiger Tail Sweep", guid: 264348 },
+  { name: "Song of Chi-Ji", guid: 198898 },
+  { name: "Ring of Peace", guid: 116844 },
+  { name: "Healing Elixir", guid: 122281 },
+  { name: "Diffuse Magic", guid: 122783 },
+  { name: "Dampen Harm", guid: 122278 },
+  { name: "Summon Jade Serpent Statue", guid: 115313 },
+  { name: "Refreshing Jade Wind", guid: 196725 },
+  { name: "Invoke Chi-Ji, the Red Crane", guid: 325197 },
+  { name: "Focused Thunder", guid: 197895 },
+  { name: "Upwelling", guid: 274963 },
+  { name: "Rising Mist", guid: 274909 },
+
+  /* -------------------------------------------------------------------------- */
+  /*                         Priest - Discipline talents                        */
+  /* -------------------------------------------------------------------------- */
+
+  { name: "Castigation", guid: 197900 },
+  { name: "Twist of Fate", guid: 265259 },
+  { name: "Schism", guid: 214621 },
+  { name: "Body and Soul", guid: 64129 },
+  { name: "Masochism", guid: 193063 },
+  { name: "Angelic Feather", guid: 121536 },
+  { name: "Shield Discipline", guid: 197045 },
+  { name: "Mindbender", guid: 123040 },
+  { name: "Power Word: Solace", guid: 129250 },
+  { name: "Psychic Voice", guid: 196704 },
+  { name: "Dominant Mind", guid: 205367 },
+  { name: "Shining Force", guid: 204263 },
+  { name: "Sins of the Many", guid: 280391 },
+  { name: "Contrition", guid: 197419 },
+  { name: "Shadow Covenant", guid: 314867 },
+  { name: "Purge the Wicked", guid: 204197 },
+  { name: "Divine Star", guid: 110744 },
+  { name: "Halo", guid: 120517 },
+  { name: "Lenience", guid: 238063 },
+  { name: "Spirit Shell", guid: 109964 },
+  { name: "Evangelism", guid: 246287 },
+
+  /* -------------------------------------------------------------------------- */
+  /*                            Priest - Holy talents                           */
+  /* -------------------------------------------------------------------------- */
+
+  { name: "Enlightenment", guid: 193155 },
+  { name: "Trail of Light", guid: 200128 },
+  { name: "Renewed Faith", guid: 341997 },
+  { name: "Angel's Mercy", guid: 238100 },
+  { name: "Body and Soul", guid: 64129 },
+  { name: "Angelic Feather", guid: 121536 },
+  { name: "Cosmic Ripple", guid: 238136 },
+  { name: "Guardian Angel", guid: 200209 },
+  { name: "Afterlife", guid: 196707 },
+  { name: "Psychic Voice", guid: 196704 },
+  { name: "Censure", guid: 200199 },
+  { name: "Shining Force", guid: 204263 },
+  { name: "Surge of Light", guid: 109186 },
+  { name: "Binding Heal", guid: 32546 },
+  { name: "Prayer Circle", guid: 321377 },
+  { name: "Benediction", guid: 193157 },
+  { name: "Divine Star", guid: 110744 },
+  { name: "Halo", guid: 120517 },
+  { name: "Light of the Naaru", guid: 196985 },
+  { name: "Apotheosis", guid: 200183 },
+  { name: "Holy Word: Salvation", guid: 265202 },
+];
