+++ src/Modules/CooldownPlanner/Data/Data.js
@@ -303,77 +303,6 @@ export const healerCooldownsDetailed = [
   },
 ];
 
-export const externalsDetailed = [
-  {
-    guid: 47788,
-    name: {
-      en: "Guardian Spirit",
-      de: "Schutzgeist",
-      fr: "Esprit gardien",
-      ru: "Оберегающий дух",
-      ch: "守护之魂",
-    },
-    wowhead: "https://www.wowhead.com/spell=47788/guardian-spirit",
-    icon: "spell_holy_guardianspirit",
-    cooldown: 180,
-  },
-
-  {
-    guid: 6940,
-    name: {
-      en: "Blessing of Sacrifice",
-      de: "Segen der Aufopferung",
-      fr: "Bénédiction de sacrifice",
-      ru: "Жертвенное благословение",
-      ch: "牺牲祝福",
-    },
-    wowhead: "https://www.wowhead.com/spell=6940/blessing-of-sacrifice",
-    icon: "spell_holy_sealofsacrifice",
-    cooldown: 120,
-  },
-  {
-    guid: 102342,
-    name: {
-      en: "Ironbark",
-      de: "Eisenborke",
-      fr: "Ecorcefer",
-      ru: "Железная кора",
-      ch: "铁木树皮",
-    },
-    wowhead: "https://www.wowhead.com/spell=102342/ironbark",
-    icon: "spell_druid_ironbark",
-    cooldown: 90,
-  },
-
-  {
-    guid: 10060,
-    name: {
-      en: "Power Infusion",
-      de: "Seele der Macht",
-      fr: "Infusion de puissance",
-      ru: "Придание сил",
-      ch: "能量灌注",
-    },
-    wowhead: "https://www.wowhead.com/spell=10060/power-infusion",
-    icon: "spell_holy_powerinfusion",
-    cooldown: 120,
-  },
-
-  {
-    guid: 116849,
-    name: {
-      en: "Life Cocoon",
-      de: "Lebenskokon",
-      fr: "Cocon de vie",
-      ru: "Исцеляющий кокон",
-      ch: "作茧缚命",
-    },
-    wowhead: "https://www.wowhead.com/spell=116849/life-cocoon",
-    icon: "ability_monk_chicocoon",
-    cooldown: 120,
-  },
-];
-
 export const bossIds = [
   {
     id: 2327,
@@ -1505,164 +1434,6 @@ export const bossAbilities = [
   },
 ];
 
-export const talents = [
-  /* -------------------------------------------------------------------------- */
-  /*                               Paladin Talents                              */
-  /* -------------------------------------------------------------------------- */
-
-  { name: "Crusader's Might", guid: 196926 },
-  { name: "Bestow Faith", guid: 223306 },
-  { name: "Light's Hammer", guid: 114158 },
-  { name: "Saved by the Light", guid: 157047 },
-  { name: "Judgment of Light", guid: 183778 },
-  { name: "Holy Prism", guid: 114165 },
-  { name: "Fist of Justice", guid: 234299 },
-  { name: "Repentance", guid: 20066 },
-  { name: "Blinding Light", guid: 115750 },
-  { name: "Unbreakable Spirit", guid: 114154 },
-  { name: "Cavalier", guid: 230332 },
-  { name: "Rule of Law", guid: 214202 },
-  { name: "Divine Purpose", guid: 223817 },
-  { name: "Holy Avenger", guid: 105809 },
-  { name: "Seraphim", guid: 152262 },
-  { name: "Sanctified Wrath", guid: 53376 },
-  { name: "Avenging Crusader", guid: 216331 },
-  { name: "Awakening", guid: 248033 },
-  { name: "Glimmer of Light", guid: 325966 },
-  { name: "Beacon of Faith", guid: 156910 },
-  { name: "Beacon of Virtue", guid: 200025 },
-
-  /* -------------------------------------------------------------------------- */
-  /*                               Shaman Talents                               */
-  /* -------------------------------------------------------------------------- */
-
-  { name: "Torrent", guid: 200072 },
-  { name: "Undulation", guid: 200071 },
-  { name: "Unleash Life", guid: 73685 },
-  { name: "Echo of the Elements", guid: 108283 },
-  { name: "Deluge", guid: 200076 },
-  { name: "Surge of Earth", guid: 320746 },
-  { name: "Spirit Wolf", guid: 260878 },
-  { name: "Earthgrab Totem", guid: 51485 },
-  { name: "Static Charge", guid: 265046 },
-  { name: "Ancestral Vigor", guid: 207401 },
-  { name: "Earthen Wall Totem", guid: 198838 },
-  { name: "Ancestral Protection Totem", guid: 207399 },
-  { name: "Nature's Guardian", guid: 30884 },
-  { name: "Graceful Spirit", guid: 192088 },
-  { name: "Wind Rush Totem", guid: 192077 },
-  { name: "Flash Flood", guid: 280614 },
-  { name: "Downpour", guid: 207778 },
-  { name: "Cloudburst Totem", guid: 157153 },
-  { name: "High Tide", guid: 157154 },
-  { name: "Wellspring", guid: 197995 },
-  { name: "Ascendance", guid: 114052 },
-
-  /* -------------------------------------------------------------------------- */
-  /*                                Druid Talents                               */
-  /* -------------------------------------------------------------------------- */
-
-  { name: "Abundance", guid: 207383 },
-  { name: "Nourish", guid: 50464 },
-  { name: "Cenarion Ward", guid: 102351 },
-  { name: "Tiger Dash", guid: 252216 },
-  { name: "Renewal", guid: 108238 },
-  { name: "Wild Charge", guid: 132302 },
-  { name: "Balance Affinity", guid: 197632 },
-  { name: "Feral Affinity", guid: 197490 },
-  { name: "Guardian Affinity", guid: 197491 },
-  { name: "Mighty Bash", guid: 5211 },
-  { name: "Mass Entanglement", guid: 102359 },
-  { name: "Heart of the Wild", guid: 319454 },
-  { name: "Soul of the Forest", guid: 158478 },
-  { name: "Cultivation", guid: 200390 },
-  { name: "Incarnation: Tree of Life", guid: 33891 },
-  { name: "Inner Peace", guid: 197073 },
-  { name: "Spring Blossoms", guid: 207385 },
-  { name: "Overgrowth", guid: 203651 },
-  { name: "Photosynthesis", guid: 274902 },
-  { name: "Germination", guid: 155675 },
-  { name: "Flourish", guid: 197721 },
-
-  /* -------------------------------------------------------------------------- */
-  /*                                Monk Talents                                */
-  /* -------------------------------------------------------------------------- */
-
-  { name: "Mist Wrap", guid: 197900 },
-  { name: "Chi Wave", guid: 115098 },
-  { name: "Chi Burst", guid: 123986 },
-  { name: "Celerity", guid: 115173 },
-  { name: "Chi Torpedo", guid: 115008 },
-  { name: "Tiger's Lust", guid: 116841 },
-  { name: "Lifecycles", guid: 197915 },
-  { name: "Spirit of the Crane", guid: 210802 },
-  { name: "Mana Tea", guid: 197908 },
-  { name: "Tiger Tail Sweep", guid: 264348 },
-  { name: "Song of Chi-Ji", guid: 198898 },
-  { name: "Ring of Peace", guid: 116844 },
-  { name: "Healing Elixir", guid: 122281 },
-  { name: "Diffuse Magic", guid: 122783 },
-  { name: "Dampen Harm", guid: 122278 },
-  { name: "Summon Jade Serpent Statue", guid: 115313 },
-  { name: "Refreshing Jade Wind", guid: 196725 },
-  { name: "Invoke Chi-Ji, the Red Crane", guid: 325197 },
-  { name: "Focused Thunder", guid: 197895 },
-  { name: "Upwelling", guid: 274963 },
-  { name: "Rising Mist", guid: 274909 },
-
-  /* -------------------------------------------------------------------------- */
-  /*                         Priest - Discipline talents                        */
-  /* -------------------------------------------------------------------------- */
-
-  { name: "Castigation", guid: 197900 },
-  { name: "Twist of Fate", guid: 265259 },
-  { name: "Schism", guid: 214621 },
-  { name: "Body and Soul", guid: 64129 },
-  { name: "Masochism", guid: 193063 },
-  { name: "Angelic Feather", guid: 121536 },
-  { name: "Shield Discipline", guid: 197045 },
-  { name: "Mindbender", guid: 123040 },
-  { name: "Power Word: Solace", guid: 129250 },
-  { name: "Psychic Voice", guid: 196704 },
-  { name: "Dominant Mind", guid: 205367 },
-  { name: "Shining Force", guid: 204263 },
-  { name: "Sins of the Many", guid: 280391 },
-  { name: "Contrition", guid: 197419 },
-  { name: "Shadow Covenant", guid: 314867 },
-  { name: "Purge the Wicked", guid: 204197 },
-  { name: "Divine Star", guid: 110744 },
-  { name: "Halo", guid: 120517 },
-  { name: "Lenience", guid: 238063 },
-  { name: "Spirit Shell", guid: 109964 },
-  { name: "Evangelism", guid: 246287 },
-
-  /* -------------------------------------------------------------------------- */
-  /*                            Priest - Holy talents                           */
-  /* -------------------------------------------------------------------------- */
-
-  { name: "Enlightenment", guid: 193155 },
-  { name: "Trail of Light", guid: 200128 },
-  { name: "Renewed Faith", guid: 341997 },
-  { name: "Angel's Mercy", guid: 238100 },
-  { name: "Body and Soul", guid: 64129 },
-  { name: "Angelic Feather", guid: 121536 },
-  { name: "Cosmic Ripple", guid: 238136 },
-  { name: "Guardian Angel", guid: 200209 },
-  { name: "Afterlife", guid: 196707 },
-  { name: "Psychic Voice", guid: 196704 },
-  { name: "Censure", guid: 200199 },
-  { name: "Shining Force", guid: 204263 },
-  { name: "Surge of Light", guid: 109186 },
-  { name: "Binding Heal", guid: 32546 },
-  { name: "Prayer Circle", guid: 321377 },
-  { name: "Benediction", guid: 193157 },
-  { name: "Divine Star", guid: 110744 },
-  { name: "Halo", guid: 120517 },
-  { name: "Light of the Naaru", guid: 196985 },
-  { name: "Apotheosis", guid: 200183 },
-  { name: "Holy Word: Salvation", guid: 265202 },
-];
-
 export const raidList = [
   // { id: 0, zoneID: 2217, raidName: "Ny’alotha" },
   { id: 0, zoneID: 2296, raidName: "Castle Nathria" },
@@ -1722,151 +1493,6 @@ export const bossList = [
     id: 2407,
     name: "Sire Denathrius",
     zoneID: 2296,
-  },
-];
-
-const spellData = [
-  // Monk Spells
-  {
-    id: 124682,
-    name: "Enveloping Mist",
-    icon: "spell_monk_envelopingmist",
-  },
-  {
-    id: 116670,
-    name: "Vivify",
-    icon: "ability_monk_vivify",
-  },
-  {
-    id: 115310,
-    name: "Revival",
-    icon: "spell_monk_revival",
-  },
-  {
-    id: 117907,
-    name: "Mastery: Gust of Mists",
-    icon: "ability_monk_souldance",
-  },
-  {
-    id: 343655,
-    name: "Enveloping Breath",
-    icon: "ability_monk_dragonkick",
-  },
-  {
-    id: 191837,
-    name: "Essence Font",
-    icon: "ability_monk_essencefont",
-  },
-  {
-    id: 115151,
-    name: "Renewing Mist",
-    icon: "ability_monk_renewingmists",
-  },
-  {
-    id: 274909,
-    name: "Rising Mist",
-    icon: "ability_monk_effuse",
-  },
-  {
-    id: 116849,
-    name: "Life Cocoon",
-    icon: "ability_monk_chicocoon",
-  },
-  {
-    id: 115175,
-    name: "Soothing Mist",
-    icon: "ability_monk_soothingmists",
-  },
-  {
-    id: 310454,
-    name: "Weapons of Order",
-    icon: "ability_bastion_monk",
-  },
-
-  // Druid Spells
-  {
-    id: 774,
-    name: "Rejuvenation",
-    icon: "spell_nature_rejuvenation",
-  },
-  {
-    id: 48438,
-    name: "Wild Growth",
-    icon: "ability_druid_flourish",
-  },
-  {
-    id: 8936,
-    name: "Regrowth",
-    icon: "spell_nature_resistnature",
-  },
-  {
-    id: 33763,
-    name: "Lifebloom",
-    icon: "inv_misc_herb_felblossom",
-  },
-
-  // Shaman Spells
-  {
-    id: 61295,
-    name: "Riptide",
-    icon: "spell_nature_riptide",
-  },
-  {
-    id: 157153,
-    name: "Cloudburst Totem",
-    icon: "ability_shaman_condensationtotem",
-  },
-  {
-    id: 73920,
-    name: "Healing Rain",
-    icon: "spell_nature_giftofthewaterspirit",
-  },
-  {
-    id: 108280,
-    name: "Healing Tide Totem",
-    icon: "ability_shaman_healingtide",
-  },
-  {
-    id: 1064,
-    name: "Chain Heal",
-    icon: "spell_nature_healingwavegreater",
-  },
-  {
-    id: 16191,
-    name: "Mana Tide Totem",
-    icon: "spell_frost_summonwaterelemental",
-  },
-  {
-    id: 8004,
-    name: "Healing Surge",
-    icon: "spell_nature_healingway",
-  },
-
-  // Holy Paladin
-  {
-    id: 225311,
-    name: "Light of Dawn",
-    icon: "spell_paladin_lightofdawn",
-  },
-  {
-    id: 85673,
-    name: "Word of Glory",
-    icon: "inv_helmet_96",
-  },
-  {
-    id: 25914,
-    name: "Holy Shock",
-    icon: "spell_holy_searinglight",
-  },
-  {
-    id: 82326,
-    name: "Holy Light",
-    icon: "spell_holy_surgeoflight",
-  },
-  {
-    id: 337824,
-    name: "Shock Barrier",
-    icon: "ability_paladin_blessedmending",
   },
 ];
 
@@ -1937,2708 +1563,7 @@ export const worldBosses = [
 
 // Boss Notes?
 // Shriekwing - Reverberating Scream consists of large damage on hit, follwoed by a dot.
-export const serverList = {
-  US: [
-    "Aegwynn",
-    "Aerie Peak",
-    "Agamaggan",
-    "Aggramar",
-    "Akama",
-    "Alexstrasza",
-    "Alleria",
-    "Altar of Storms",
-    "Alterac Mountains",
-    "Aman'Thul",
-    "Andorhal",
-    "Anetheron",
-    "Antonidas",
-    "Anub'arak",
-    "Anvilmar",
-    "Arathor",
-    "Archimonde",
-    "Area 52",
-    "Argent Dawn",
-    "Arthas",
-    "Arygos",
-    "Auchindoun",
-    "Azgalor",
-    "Azjol-Nerub",
-    "Azralon",
-    "Azshara",
-    "Azuremyst",
-    "Baelgun",
-    "Balnazzar",
-    "Barthilas",
-    "Black Dragonflight",
-    "Blackhand",
-    "Blackrock",
-    "Blackwater Raiders",
-    "Blackwing Lair",
-    "Blade's Edge",
-    "Bladefist",
-    "Bleeding Hollow",
-    "Blood Furnace",
-    "Bloodhoof",
-    "Bloodscalp",
-    "Bonechewer",
-    "Borean Tundra",
-    "Boulderfist",
-    "Bronzebeard",
-    "Burning Blade",
-    "Burning Legion",
-    "Caelestrasz",
-    "Cairne",
-    "Cenarion Circle",
-    "Cenarius",
-    "Cho'gall",
-    "Chromaggus",
-    "Coilfang",
-    "Crushridge",
-    "Daggerspine",
-    "Dalaran",
-    "Dalvengyr",
-    "Dark Iron",
-    "Darkspear",
-    "Darrowmere",
-    "Dath'Remar",
-    "Dawnbringer",
-    "Deathwing",
-    "Demon Soul",
-    "Dentarg",
-    "Destromath",
-    "Dethecus",
-    "Detheroc",
-    "Doomhammer",
-    "Draenor",
-    "Dragonblight",
-    "Dragonmaw",
-    "Drak'Tharon",
-    "Drak'thul",
-    "Draka",
-    "Drakkari",
-    "Dreadmaul",
-    "Drenden",
-    "Dunemaul",
-    "Durotan",
-    "Duskwood",
-    "Earthen Ring",
-    "Echo Isles",
-    "Eitrigg",
-    "Eldre'Thalas",
-    "Elune",
-    "Emerald Dream",
-    "Eonar",
-    "Eredar",
-    "Executus",
-    "Exodar",
-    "Farstriders",
-    "Feathermoon",
-    "Fenris",
-    "Firetree",
-    "Fizzcrank",
-    "Frostmane",
-    "Frostmourne",
-    "Frostwolf",
-    "Galakrond",
-    "Gallywix",
-    "Garithos",
-    "Garona",
-    "Garrosh",
-    "Ghostlands",
-    "Gilneas",
-    "Gnomeregan",
-    "Goldrinn",
-    "Gorefiend",
-    "Gorgonnash",
-    "Greymane",
-    "Grizzly Hills",
-    "Gul'dan",
-    "Gundrak",
-    "Gurubashi",
-    "Hakkar",
-    "Haomarush",
-    "Hellscream",
-    "Hydraxis",
-    "Hyjal",
-    "Icecrown",
-    "Illidan",
-    "Jaedenar",
-    "Jubei'Thos",
-    "Kael'thas",
-    "Kalecgos",
-    "Kargath",
-    "Kel'Thuzad",
-    "Khadgar",
-    "Khaz Modan",
-    "Khaz'goroth",
-    "Kil'jaeden",
-    "Kilrogg",
-    "Kirin Tor",
-    "Korgath",
-    "Korialstrasz",
-    "Kul Tiras",
-    "Laughing Skull",
-    "Lethon",
-    "Lightbringer",
-    "Lightning's Blade",
-    "Lightninghoof",
-    "Llane",
-    "Lothar",
-    "Madoran",
-    "Maelstrom",
-    "Magtheridon",
-    "Maiev",
-    "Mal'Ganis",
-    "Malfurion",
-    "Malorne",
-    "Malygos",
-    "Mannoroth",
-    "Medivh",
-    "Misha",
-    "Mok'Nathal",
-    "Moon Guard",
-    "Moonrunner",
-    "Mug'thol",
-    "Muradin",
-    "Nagrand",
-    "Nathrezim",
-    "Nazgrel",
-    "Nazjatar",
-    "Nemesis",
-    "Ner'zhul",
-    "Nesingwary",
-    "Nordrassil",
-    "Norgannon",
-    "Onyxia",
-    "Perenolde",
-    "Proudmoore",
-    "Quel'Thalas",
-    "Quel'dorei",
-    "Ragnaros",
-    "Ravencrest",
-    "Ravenholdt",
-    "Rexxar",
-    "Rivendare",
-    "Runetotem",
-    "Sargeras",
-    "Saurfang",
-    "Scarlet Crusade",
-    "Scilla",
-    "Sen'jin",
-    "Sentinels",
-    "Shadow Council",
-    "Shadowmoon",
-    "Shadowsong",
-    "Shandris",
-    "Shattered Halls",
-    "Shattered Hand",
-    "Shu'halo",
-    "Silver Hand",
-    "Silvermoon",
-    "Sisters of Elune",
-    "Skullcrusher",
-    "Skywall",
-    "Smolderthorn",
-    "Spinebreaker",
-    "Spirestone",
-    "Staghelm",
-    "Steamwheedle Cartel",
-    "Stonemaul",
-    "Stormrage",
-    "Stormreaver",
-    "Stormscale",
-    "Suramar",
-    "Tanaris",
-    "Terenas",
-    "Terokkar",
-    "Thaurissan",
-    "The Forgotten Coast",
-    "The Scryers",
-    "The Underbog",
-    "The Venture Co",
-    "Thorium Brotherhood",
-    "Thrall",
-    "Thunderhorn",
-    "Thunderlord",
-    "Tichondrius",
-    "Tol Barad",
-    "Tortheldrin",
-    "Trollbane",
-    "Turalyon",
-    "Twisting Nether",
-    "Uldaman",
-    "Uldum",
-    "Undermine",
-    "Ursin",
-    "Uther",
-    "Vashj",
-    "Vek'nilash",
-    "Velen",
-    "Warsong",
-    "Whisperwind",
-    "Wildhammer",
-    "Windrunner",
-    "Winterhoof",
-    "Wyrmrest Accord",
-    "Ysera",
-    "Ysondre",
-    "Zangarmarsh",
-    "Zul'jin",
-    "Zuluhed",
-  ],
-  EU: [
-    "Aegwynn",
-    "Aerie Peak",
-    "Agamaggan",
-    "Aggra (Português)",
-    "Aggramar",
-    "Ahn'Qiraj",
-    "Al'Akir",
-    "Alexstrasza",
-    "Alleria",
-    "Alonsus",
-    "Aman'Thul",
-    "Ambossar",
-    "Anachronos",
-    "Anetheron",
-    "Antonidas",
-    "Anub'arak",
-    "Arak-arahm",
-    "Arathi",
-    "Arathor",
-    "Archimonde",
-    "Area 52",
-    "Argent Dawn",
-    "Arthas",
-    "Arygos",
-    "Ashenvale",
-    "Aszune",
-    "Auchindoun",
-    "Azjol-Nerub",
-    "Azshara",
-    "Azuregos",
-    "Azuremyst",
-    "Baelgun",
-    "Balnazzar",
-    "Blackhand",
-    "Blackmoore",
-    "Blackrock",
-    "Blackscar",
-    "Blade's Edge",
-    "Bladefist",
-    "Bloodfeather",
-    "Bloodhoof",
-    "Bloodscalp",
-    "Blutkessel",
-    "Booty Bay",
-    "Borean Tundra",
-    "Boulderfist",
-    "Bronze Dragonflight",
-    "Bronzebeard",
-    "Burning Blade",
-    "Burning Legion",
-    "Burning Steppes",
-    "C'Thun",
-    "Chamber of Aspects",
-    "Chants éternels",
-    "Cho'gall",
-    "Chromaggus",
-    "Colinas Pardas",
-    "Confrérie du Thorium",
-    "Conseil des Ombres",
-    "Crushridge",
-    "Culte de la Rive noire",
-    "Daggerspine",
-    "Dalaran",
-    "Dalvengyr",
-    "Darkmoon Faire",
-    "Darksorrow",
-    "Darkspear",
-    "Das Konsortium",
-    "Das Syndikat",
-    "Deathguard",
-    "Deathweaver",
-    "Deathwing",
-    "Deepholm",
-    "Defias Brotherhood",
-    "Dentarg",
-    "Der Mithrilorden",
-    "Der Rat von Dalaran",
-    "Der abyssische Rat",
-    "Destromath",
-    "Dethecus",
-    "Die Aldor",
-    "Die Arguswacht",
-    "Die Nachtwache",
-    "Die Silberne Hand",
-    "Die Todeskrallen",
-    "Die ewige Wacht",
-    "Doomhammer",
-    "Draenor",
-    "Dragonblight",
-    "Dragonmaw",
-    "Drak'thul",
-    "Drek'Thar",
-    "Dun Modr",
-    "Dun Morogh",
-    "Dunemaul",
-    "Durotan",
-    "Earthen Ring",
-    "Echsenkessel",
-    "Eitrigg",
-    "Eldre'Thalas",
-    "Elune",
-    "Emerald Dream",
-    "Emeriss",
-    "Eonar",
-    "Eredar",
-    "Eversong",
-    "Executus",
-    "Exodar",
-    "Festung der Stürme",
-    "Fordragon",
-    "Forscherliga",
-    "Frostmane",
-    "Frostmourne",
-    "Frostwhisper",
-    "Frostwolf",
-    "Galakrond",
-    "Garona",
-    "Garrosh",
-    "Genjuros",
-    "Ghostlands",
-    "Gilneas",
-    "Goldrinn",
-    "Gordunni",
-    "Gorgonnash",
-    "Greymane",
-    "Grim Batol",
-    "Grom",
-    "Gul'dan",
-    "Hakkar",
-    "Haomarush",
-    "Hellfire",
-    "Hellscream",
-    "Howling Fjord",
-    "Hyjal",
-    "Illidan",
-    "Jaedenar",
-    "Kael'thas",
-    "Karazhan",
-    "Kargath",
-    "Kazzak",
-    "Kel'Thuzad",
-    "Khadgar",
-    "Khaz Modan",
-    "Khaz'goroth",
-    "Kil'jaeden",
-    "Kilrogg",
-    "Kirin Tor",
-    "Kor'gall",
-    "Krag'jin",
-    "Krasus",
-    "Kul Tiras",
-    "Kult der Verdammten",
-    "La Croisade écarlate",
-    "Laughing Skull",
-    "Les Clairvoyants",
-    "Les Sentinelles",
-    "Lich King",
-    "Lightbringer",
-    "Lightning's Blade",
-    "Lordaeron",
-    "Los Errantes",
-    "Lothar",
-    "Madmortem",
-    "Magtheridon",
-    "Mal'Ganis",
-    "Malfurion",
-    "Malorne",
-    "Malygos",
-    "Mannoroth",
-    "Marécage de Zangar",
-    "Mazrigos",
-    "Medivh",
-    "Minahonda",
-    "Moonglade",
-    "Mug'thol",
-    "Nagrand",
-    "Nathrezim",
-    "Naxxramas",
-    "Nazjatar",
-    "Nefarian",
-    "Nemesis",
-    "Neptulon",
-    "Ner'zhul",
-    "Nera'thor",
-    "Nethersturm",
-    "Nordrassil",
-    "Norgannon",
-    "Nozdormu",
-    "Onyxia",
-    "Outland",
-    "Perenolde",
-    "Pozzo dell'Eternità",
-    "Proudmoore",
-    "Quel'Thalas",
-    "Ragnaros",
-    "Rajaxx",
-    "Rashgarroth",
-    "Ravencrest",
-    "Ravenholdt",
-    "Razuvious",
-    "Rexxar",
-    "Runetotem",
-    "Sanguino",
-    "Sargeras",
-    "Saurfang",
-    "Scarshield Legion",
-    "Sen'jin",
-    "Shadowsong",
-    "Shattered Halls",
-    "Shattered Hand",
-    "Shattrath",
-    "Shen'dralar",
-    "Silvermoon",
-    "Sinstralis",
-    "Skullcrusher",
-    "Soulflayer",
-    "Spinebreaker",
-    "Sporeggar",
-    "Steamwheedle Cartel",
-    "Stormrage",
-    "Stormreaver",
-    "Stormscale",
-    "Sunstrider",
-    "Suramar",
-    "Sylvanas",
-    "Taerar",
-    "Talnivarr",
-    "Tarren Mill",
-    "Teldrassil",
-    "Temple noir",
-    "Terenas",
-    "Terokkar",
-    "Terrordar",
-    "The Maelstrom",
-    "The Sha'tar",
-    "The Venture Co",
-    "Theradras",
-    "Thermaplugg",
-    "Thrall",
-    "Throk'Feroth",
-    "Thunderhorn",
-    "Tichondrius",
-    "Tirion",
-    "Todeswache",
-    "Trollbane",
-    "Turalyon",
-    "Twilight's Hammer",
-    "Twisting Nether",
-    "Tyrande",
-    "Uldaman",
-    "Ulduar",
-    "Uldum",
-    "Un'Goro",
-    "Varimathras",
-    "Vashj",
-    "Vek'lor",
-    "Vek'nilash",
-    "Vol'jin",
-    "Wildhammer",
-    "Wrathbringer",
-    "Xavius",
-    "Ysera",
-    "Ysondre",
-    "Zenedar",
-    "Zirkel des Cenarius",
-    "Zul'jin",
-    "Zuluhed",
-  ],
-  KR: [
-    "Alexstrasza",
-    "Azshara",
-    "Burning Legion",
-    "Cenarius",
-    "Dalaran",
-    "Deathwing",
-    "Durotan",
-    "Garona",
-    "Gul'dan",
-    "Hellscream",
-    "Hyjal",
-    "Malfurion",
-    "Norgannon",
-    "Rexxar",
-    "Stormrage",
-    "Wildhammer",
-    "Windrunner",
-    "Zul'jin",
-  ],
-  TW: [
-    "Arthas",
-    "Arygos",
-    "Bleeding Hollow",
-    "Chillwind Point",
-    "Crystalpine Stinger",
-    "Demon Fall Canyon",
-    "Dragonmaw",
-    "Frostmane",
-    "Hellscream",
-    "Icecrown",
-    "Light's Hope",
-    "Menethil",
-    "Nightsong",
-    "Order of the Cloud Serpent",
-    "Quel'dorei",
-    "Shadowmoon",
-    "Silverwing Hold",
-    "Skywall",
-    "Spirestone",
-    "Stormscale",
-    "Sundown Marsh",
-    "Whisperwind",
-    "World Tree",
-    "Wrathbringer",
-    "Zealot Blade",
-  ],
-  CN: [
-    "万色星辰",
-    "世界之树",
-    "丹莫德",
-    "主宰之剑",
-    "丽丽（四川",
-    "亚雷戈斯",
-    "亡语者",
-    "伊兰尼库斯",
-    "伊利丹",
-    "伊森利恩",
-    "伊森德雷",
-    "伊瑟拉",
-    "伊莫塔尔",
-    "伊萨里奥斯",
-    "元素之力",
-    "克尔苏加德",
-    "克洛玛古斯",
-    "克苏恩",
-    "兰娜瑟尔",
-    "军团要塞",
-    "冬寒",
-    "冬拥湖",
-    "冬泉谷",
-    "冰川之拳",
-    "冰霜之刃",
-    "冰风岗",
-    "凤凰之神",
-    "凯尔萨斯",
-    "凯恩血蹄",
-    "刀塔",
-    "利刃之拳",
-    "刺骨利刃",
-    "加兹鲁维",
-    "加基森",
-    "加尔",
-    "加里索斯",
-    "勇士岛",
-    "千针石林",
-    "卡德加",
-    "卡德罗斯",
-    "卡扎克",
-    "卡拉赞",
-    "卡珊德拉",
-    "厄祖玛特",
-    "双子峰",
-    "古加尔",
-    "古尔丹",
-    "古拉巴什",
-    "古达克",
-    "哈兰",
-    "哈卡",
-    "嚎风峡湾",
-    "回音山",
-    "国王之谷",
-    "图拉扬",
-    "圣火神殿",
-    "地狱之石",
-    "地狱咆哮",
-    "埃克索图斯",
-    "埃加洛尔",
-    "埃基尔松",
-    "埃德萨拉",
-    "埃苏雷格",
-    "埃雷达尔",
-    "埃霍恩",
-    "基尔加丹",
-    "基尔罗格",
-    "塔伦米尔",
-    "塔纳利斯",
-    "塞拉摩",
-    "塞拉赞恩",
-    "塞泰克",
-    "塞纳留斯",
-    "壁炉谷",
-    "夏维安",
-    "外域",
-    "大地之怒",
-    "大漩涡",
-    "天空之墙",
-    "天谴之门",
-    "太阳之井",
-    "夺灵者",
-    "奈法利安",
-    "奈萨里奥",
-    "奎尔丹纳斯",
-    "奎尔萨拉斯",
-    "奥妮克希亚",
-    "奥尔加隆",
-    "奥拉基尔",
-    "奥斯里安",
-    "奥杜尔",
-    "奥特兰克",
-    "奥蕾莉亚",
-    "奥达曼",
-    "奥金顿",
-    "守护之剑",
-    "安东尼达斯",
-    "安其拉",
-    "安加萨",
-    "安威玛尔",
-    "安戈洛",
-    "安格博达",
-    "安纳塞隆",
-    "安苏",
-    "密林游侠",
-    "寒冰皇冠",
-    "尘风峡谷",
-    "屠魔山谷",
-    "山丘之王",
-    "巨龙之吼",
-    "巫妖之王",
-    "巴尔古恩",
-    "巴瑟拉斯",
-    "巴纳扎尔",
-    "布兰卡德",
-    "布莱克摩",
-    "布莱恩",
-    "布鲁塔卢斯",
-    "希尔瓦娜斯",
-    "希雷诺斯",
-    "幽暗沼泽",
-    "库尔提拉斯",
-    "库德兰",
-    "弗塞雷迦",
-    "影之哀伤",
-    "影牙要塞",
-    "德拉诺",
-    "恐怖图腾",
-    "恶魔之翼",
-    "恶魔之魂",
-    "戈古纳斯",
-    "戈提克",
-    "战歌",
-    "扎拉赞恩",
-    "托塞德林",
-    "托尔巴拉德",
-    "拉文凯斯",
-    "拉文霍德",
-    "拉格纳罗斯",
-    "拉贾克斯",
-    "提尔之手",
-    "提瑞斯法",
-    "摩摩尔",
-    "斩魔者",
-    "斯克提斯",
-    "斯坦索姆",
-    "无尽之海",
-    "无底海渊",
-    "日落沼泽",
-    "时光之穴",
-    "普瑞斯托",
-    "普罗德摩",
-    "晴日峰（江苏）",
-    "暗影之月",
-    "暗影裂口",
-    "暗影议会",
-    "暗影迷宫",
-    "暮色森林",
-    "暴风祭坛",
-    "月光林地",
-    "月神殿",
-    "末日祷告祭坛",
-    "末日行者",
-    "朵丹尼尔",
-    "杜隆坦",
-    "格瑞姆巴托",
-    "格雷迈恩",
-    "格鲁尔",
-    "桑德兰",
-    "梅尔加尼",
-    "梦境之树",
-    "森金",
-    "死亡之翼",
-    "死亡熔炉",
-    "毁灭之锤",
-    "永夜港",
-    "永恒之井",
-    "沃金",
-    "沙怒",
-    "法拉希姆",
-    "泰兰德",
-    "泰拉尔",
-    "洛丹伦",
-    "洛肯",
-    "洛萨",
-    "海克泰尔",
-    "海加尔",
-    "海达希亚",
-    "深渊之喉",
-    "深渊之巢",
-    "激流之傲",
-    "激流堡",
-    "火喉",
-    "火烟之谷",
-    "火焰之树",
-    "火羽山",
-    "灰谷",
-    "烈焰峰",
-    "烈焰荆棘",
-    "熊猫酒仙",
-    "熔火之心",
-    "熵魔",
-    "燃烧之刃",
-    "燃烧军团",
-    "燃烧平原",
-    "爱斯特纳",
-    "狂热之刃",
-    "狂风峭壁",
-    "玛多兰",
-    "玛法里奥",
-    "玛洛加尔",
-    "玛瑟里顿",
-    "玛维·影歌",
-    "玛诺洛斯",
-    "玛里苟斯",
-    "瑞文戴尔",
-    "瑟莱德丝",
-    "瓦丝琪",
-    "瓦拉斯塔兹",
-    "瓦拉纳",
-    "瓦里安",
-    "瓦里玛萨斯",
-    "甜水绿洲",
-    "生态船",
-    "白银之手",
-    "白骨荒野",
-    "盖斯",
-    "石爪峰",
-    "石锤",
-    "破碎岭",
-    "祖尔金",
-    "祖达克",
-    "祖阿曼",
-    "神圣之歌",
-    "穆戈尔",
-    "符文图腾",
-    "米奈希尔",
-    "索拉丁",
-    "索瑞森",
-    "红云台地",
-    "红龙军团",
-    "红龙女王",
-    "纳克萨玛斯",
-    "纳沙塔尔",
-    "织亡者",
-    "罗宁",
-    "罗曼斯",
-    "羽月",
-    "翡翠梦境",
-    "耐奥祖",
-    "耐普图隆",
-    "耳语海岸",
-    "能源舰",
-    "自由之风",
-    "艾森娜",
-    "艾欧娜尔",
-    "艾维娜",
-    "艾苏恩",
-    "艾莫莉丝",
-    "艾萨拉",
-    "艾露恩",
-    "芬里斯",
-    "苏塔恩",
-    "苏拉玛",
-    "范克里夫",
-    "范达尔鹿盔",
-    "荆棘谷",
-    "莱索恩",
-    "菲拉斯",
-    "菲米丝",
-    "萨尔",
-    "萨格拉斯",
-    "萨洛拉丝",
-    "萨菲隆",
-    "蓝龙军团",
-    "藏宝海湾",
-    "蜘蛛王国",
-    "血吼",
-    "血牙魔王",
-    "血环",
-    "血羽",
-    "血色十字军",
-    "血顶",
-    "试炼之环",
-    "诺兹多姆",
-    "诺森德",
-    "诺莫瑞根",
-    "贫瘠之地",
-    "踏梦者",
-    "轻风之语",
-    "辛达苟萨",
-    "达克萨隆",
-    "达基萨斯",
-    "达尔坎",
-    "达文格尔",
-    "达斯雷玛",
-    "达纳斯",
-    "达隆米尔",
-    "迅捷微风",
-    "远古海滩",
-    "迦拉克隆",
-    "迦玛兰",
-    "迦罗娜",
-    "迦顿",
-    "迪托马斯",
-    "迪瑟洛克",
-    "逐日者",
-    "通灵学院",
-    "遗忘海岸",
-    "金度",
-    "金色平原",
-    "铜龙军团",
-    "银月",
-    "银松森林",
-    "闪电之刃",
-    "阿克蒙德",
-    "阿努巴拉克",
-    "阿卡玛",
-    "阿古斯",
-    "阿尔萨斯",
-    "阿扎达斯",
-    "阿拉希",
-    "阿拉索",
-    "阿斯塔洛",
-    "阿曼尼",
-    "阿格拉玛",
-    "阿比迪斯",
-    "阿纳克洛斯",
-    "阿迦玛甘",
-    "雏龙之翼",
-    "雷克萨",
-    "雷斧堡垒",
-    "雷霆之怒",
-    "雷霆之王",
-    "雷霆号角",
-    "霍格",
-    "霜之哀伤",
-    "霜狼",
-    "风暴之怒",
-    "风暴之眼",
-    "风暴之鳞",
-    "风暴峭壁",
-    "风行者",
-    "鬼雾峰",
-    "鲜血熔炉",
-    "鹰巢山",
-    "麦姆",
-    "麦迪文",
-    "黄金之路",
-    "黑手军团",
-    "黑暗之矛",
-    "黑暗之门",
-    "黑暗虚空",
-    "黑暗魅影",
-    "黑石尖塔",
-    "黑翼之巢",
-    "黑铁",
-    "黑锋哨站",
-    "黑龙军团",
-    "龙骨平原",
-  ],
-};
 
-export const conduitDB = [
-  {
-    guid: 0,
-    icon: "inv_axe_02",
-    name: "Unknown",
-  },
-
-  /* -------------------------------------------------------------------------- */
-  /*                                    Druid                                   */
-  /* -------------------------------------------------------------------------- */
-
-  /* ---------------------------- Potency Conduits ---------------------------- */
-
-  {
-    guid: 340616,
-    abilityIcon: "spell_nature_crystalball.jpg",
-    name: {
-      en: "Flash of Clarity",
-      cn: "明晰闪愈",
-      de: "Augenblick der Klarsicht",
-      fr: "Instant de lucidité",
-      ru: "Вспышка ясности",
-    },
-    type: "Potency",
-    class: "Restoration Druid",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 340621,
-    abilityIcon: "item_herbd.jpg",
-    name: {
-      en: "Floral Recycling",
-      cn: "花开花谢",
-      de: "Pflanzliche Wiederherstellung",
-      fr: "Recyclage floral",
-      ru: "Переработка цветов",
-    },
-    type: "Potency",
-    class: "Restoration Druid",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 340550,
-    abilityIcon: "spell_nature_ravenform.jpg",
-    name: {
-      en: "Ready for Anything",
-      cn: "准备万全",
-      de: "Auf alles gefasst",
-      fr: "Prêt à tout",
-      ru: "Готовность ко всему",
-    },
-    type: "Potency",
-    class: "Restoration Druid",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 340549,
-    abilityIcon: "ability_druid_flourish.jpg",
-    name: {
-      en: "Unstoppable Growth",
-      cn: "狂野生长",
-      de: "Unaufhaltsames Wachstum",
-      fr: "Croissance irrésistible",
-      ru: "Беспрестанный рост",
-    },
-    type: "Potency",
-    class: "Restoration Druid",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 341446,
-    abilityIcon: "ability_ardenweald_druid.jpg",
-    name: {
-      en: "Conflux of Elements",
-      cn: "元素汇流",
-      de: "Konflux der Elemente",
-      fr: "Confluence des éléments",
-      ru: "Слияние стихий",
-    },
-    type: "Potency",
-    class: "Restoration Druid",
-    covenant: "Night Fae",
-    show: true,
-  },
-  {
-    guid: 341378,
-    abilityIcon: "ability_bastion_druid.jpg",
-    name: {
-      en: "Deep Allegiance",
-      cn: "诚挚忠心",
-      de: "Tiefe Loyalität",
-      fr: "Allégeance indéfectible",
-      ru: "Беспредельная верность",
-    },
-    type: "Potency",
-    class: "Restoration Druid",
-    covenant: "Kyrian",
-    show: true,
-  },
-  {
-    guid: 341383,
-    abilityIcon: "ability_revendreth_druid.jpg",
-    name: {
-      en: "Endless Thirst",
-      cn: "无穷饥渴",
-      de: "Endloser Durst",
-      fr: "Soif éternelle",
-      ru: "Неутолимая жажда",
-    },
-    type: "Potency",
-    class: "Restoration Druid",
-    covenant: "Venthyr",
-    show: true,
-  },
-  {
-    guid: 341447,
-    abilityIcon: "ability_maldraxxus_druid.jpg",
-    name: {
-      en: "Evolved Swarm",
-      cn: "进化蜂群",
-      de: "Evolvierter Schwarm",
-      fr: "Essaim évolué",
-      ru: "Эволюционировавший рой",
-    },
-    type: "Potency",
-    class: "Restoration Druid",
-    covenant: "Necrolord",
-    show: true,
-  },
-
-  /* --------------------------- Endurance Conduits --------------------------- */
-
-  {
-    guid: 340529,
-    abilityIcon: "spell_nature_stoneclawtotem.jpg",
-    name: {
-      en: "Tough as Bark",
-      cn: "韧如树皮",
-      de: "Borkenhart",
-      fr: "Dur comme l’écorce",
-      ru: "Древесная жесткость",
-    },
-    type: "Endurance",
-    class: "Restoration Druid",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 340553,
-    abilityIcon: "ability_druid_tigersroar.jpg",
-    name: {
-      en: "Well-Honed Instincts",
-      cn: "精确本能",
-      de: "Geschärfte Instinkte",
-      fr: "Instinct infaillible",
-      ru: "Отточенные инстинкты",
-    },
-    type: "Endurance",
-    class: "Restoration Druid",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 340540,
-    abilityIcon: "ability_druid_markofursol.jpg",
-    name: {
-      en: "Ursine Vigor",
-      cn: "巨熊活力",
-      de: "Kraft des Bären",
-      fr: "Vigueur ursine",
-      ru: "Медвежья энергия",
-    },
-    type: "Endurance",
-    class: "Restoration Druid",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 340543,
-    abilityIcon: "spell_nature_healingway.jpg",
-    name: {
-      en: "Innate Resolve",
-      cn: "巨熊固有决心活力",
-      de: "Angeborene Entschlossenheit",
-      fr: "Résolution innée",
-      ru: "Врожденная стойкость",
-    },
-    type: "Endurance",
-    class: "Restoration Druid",
-    covenant: "",
-    show: true,
-  },
-
-  /* ---------------------------- Finesse Conduits ---------------------------- */
-
-  {
-    guid: 341280,
-    abilityIcon: "spell_nature_reincarnation.jpg",
-    name: {
-      en: "Born Anew",
-      cn: "重获新生",
-      de: "Neu geboren",
-      fr: "Seconde naissance",
-      ru: "Рождение заново",
-    },
-    type: "Finesse",
-    class: "Restoration Druid",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 341450,
-    abilityIcon: "spell_druid_stampedingroar_cat.jpg",
-    name: {
-      en: "Front of the Pack",
-      cn: "兽群先锋",
-      de: "Spitze des Rudels",
-      fr: "Tête de la meute",
-      ru: "Во главе стаи",
-    },
-    type: "Finesse",
-    class: "Restoration Druid",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 341451,
-    abilityIcon: "ability_hunter_huntervswild.jpg",
-    name: {
-      en: "Born of the Wilds",
-      cn: "荒野之子",
-      de: "Kind der Wildnis",
-      fr: "Enfant sauvage",
-      ru: "Дитя дикой природы",
-    },
-    type: "Finesse",
-    class: "Restoration Druid",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 340545,
-    abilityIcon: "spell_druid_tirelesspursuit.jpg",
-    name: {
-      en: "Tireless Pursuit",
-      cn: "不息追猎",
-      de: "Unermüdliche Verfolgung",
-      fr: "Poursuite inlassable",
-      ru: "Неустанная погоня",
-    },
-    type: "Finesse",
-    class: "Restoration Druid",
-    covenant: "",
-    show: true,
-  },
-
-  /* -------------------------------------------------------------------------- */
-  /*                                    Monk                                    */
-  /* -------------------------------------------------------------------------- */
-
-  /* ---------------------------- Potency Conduits ---------------------------- */
-
-  {
-    guid: 336773,
-    abilityIcon: "inv_inscription_deck_jadeserpent.jpg",
-    name: {
-      en: "Jade Bond",
-      cn: "翡翠羁绊",
-      de: "Jadeband",
-      fr: "Lien de jade",
-      ru: "Связь c Нефритовой змеей",
-    },
-    type: "Potency",
-    class: "Mistweaver Monk",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 337241,
-    abilityIcon: "inv_misc_gem_pearl_06.jpg",
-    name: {
-      en: "Nourishing Chi",
-      cn: "滋养真气",
-      de: "Nährendes Chi",
-      fr: "Chi nourrissant",
-      ru: "Укрепляющая ци",
-    },
-    type: "Potency",
-    class: "Mistweaver Monk",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 337099,
-    abilityIcon: "spell_monk_revival.jpg",
-    name: {
-      en: "Rising Sun Revival",
-      cn: "旭日还魂",
-      de: "Belebung der aufgehenden Sonne",
-      fr: "Renaissance du soleil levant",
-      ru: "Восстановление восходящего солнца",
-    },
-    type: "Potency",
-    class: "Mistweaver Monk",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 336812,
-    abilityIcon: "spell_nature_abolishmagic.jpg",
-    name: {
-      en: "Resplendent Mist",
-      cn: "璀璨迷雾",
-      de: "Prächtiger Nebel",
-      fr: "Brume resplendissante",
-      ru: "Сверкающий туман",
-    },
-    type: "Potency",
-    class: "Mistweaver Monk",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 337286,
-    abilityIcon: "ability_bastion_warrior.jpg",
-    name: {
-      en: "Strike with Clarity",
-      cn: "洞察一击",
-      de: "Schlag voller Klarheit",
-      fr: "Frappe de clarté",
-      ru: "Удар просветленных",
-    },
-    type: "Potency",
-    class: "Mistweaver Monk",
-    covenant: "Kyrian",
-    show: true,
-  },
-  {
-    guid: 337301,
-    abilityIcon: "spell_animarevendreth_buff.jpg",
-    name: {
-      en: "Imbued Reflections",
-      cn: "注能映像",
-      de: "Verzauberte Reflexionen",
-      fr: "Reflets imprégnés",
-      ru: "Усиленные отражения",
-    },
-    type: "Potency",
-    class: "Mistweaver Monk",
-    covenant: "Venthyr",
-    show: true,
-  },
-  {
-    guid: 337295,
-    abilityIcon: "spell_animamaldraxxus_nova.jpg",
-    name: {
-      en: "Bone Marrow Hops",
-      cn: "骨髓酒花",
-      de: "Knochenmarkhopfen",
-      fr: "Houblon à moelle",
-      ru: "Хмель из костного мозга",
-    },
-    type: "Potency",
-    class: "Mistweaver Monk",
-    covenant: "Necrolords",
-    show: true,
-  },
-  {
-    guid: 337303,
-    abilityIcon: "spell_animaardenweald_beam.jpg",
-    name: {
-      en: "Way of the Fae",
-      cn: "法夜之道",
-      de: "Weg der Fae",
-      fr: "Voie des Faë",
-      ru: "Путь ночного народца",
-    },
-    type: "Potency",
-    class: "Mistweaver Monk",
-    covenant: "Night Fae",
-    show: true,
-  },
-
-  /* --------------------------- Endurance Conduits --------------------------- */
-
-  {
-    guid: 336853,
-    abilityIcon: "inv_misc_food_vendor_blackpepper.jpg",
-    name: {
-      en: "Fortifying Ingredients",
-      cn: "壮胆配料",
-      de: "Stärkende Zutaten",
-      fr: "Ingrédients fortifiants",
-      ru: "Укрепляющие ингредиенты",
-    },
-    type: "Endurance",
-    class: "Mistweaver Monk",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 336632,
-    abilityIcon: "inv_misc_volatileearth.jpg",
-    name: {
-      en: "Grounding Breath",
-      cn: "定息活血",
-      de: "Erdender Atem",
-      fr: "Souffle isolant",
-      ru: "Заземляющее дыхание",
-    },
-    type: "Endurance",
-    class: "Mistweaver Monk",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 336379,
-    abilityIcon: "ability_monk_chiswirl.jpg",
-    name: {
-      en: "Harm Denial",
-      cn: "接木成林",
-      de: "Schadensabwendung",
-      fr: "Refus du mal",
-      ru: "Защита от вреда",
-    },
-    type: "Endurance",
-    class: "Mistweaver Monk",
-    covenant: "",
-    show: true,
-  },
-
-  /* ---------------------------- Finesse Conduits ---------------------------- */
-
-  {
-    guid: 336890,
-    abilityIcon: "inv_elemental_mote_air01.jpg",
-    name: {
-      en: "Dizzying Tumble",
-      cn: "失神倾倒",
-      de: "Schwindelerregende Akrobatik",
-      fr: "Chute étourdissante",
-      ru: "Ошеломляющий кувырок",
-    },
-    type: "Finesse",
-    class: "Mistweaver Monk",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 336884,
-    abilityIcon: "sha_ability_rogue_bloodyeye.jpg",
-    name: {
-      en: "Lingering Numbness",
-      cn: "骨软筋酥",
-      de: "Verweilende Taubheit",
-      fr: "Engourdissement persistant",
-      ru: "Остаточное оцепенение",
-    },
-    type: "Finesse",
-    class: "Mistweaver Monk",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 337078,
-    abilityIcon: "ability_monk_ridethewind.jpg",
-    name: {
-      en: "Swift Transference",
-      cn: "迅捷双分",
-      de: "Flinker Transfer",
-      fr: "Transfert rapide",
-      ru: "Стремительное перенаправление",
-    },
-    type: "Finesse",
-    class: "Mistweaver Monk",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 337084,
-    abilityIcon: "ability_titankeeper_phasing.jpg",
-    name: {
-      en: "Tumbling Technique",
-      cn: "身法决",
-      de: "Akrobatisches Abrollen",
-      fr: "Art de la chute",
-      ru: "Техника кувырка",
-    },
-    type: "Finesse",
-    class: "Mistweaver Monk",
-    covenant: "",
-    show: true,
-  },
-
-  /* -------------------------------------------------------------------------- */
-  /*                                   Paladin                                  */
-  /* -------------------------------------------------------------------------- */
-
-  /* ---------------------------- Potency Conduits ---------------------------- */
-
-  {
-    guid: 339570,
-    abilityIcon: "ability_priest_spiritoftheredeemer.jpg",
-    name: {
-      en: "Enkindled Spirit",
-      cn: "光明之魂",
-      de: "Entflammter Geist",
-      fr: "Esprit embrasé",
-      ru: "Воодушевленный дух",
-    },
-    type: "Potency",
-    class: "Holy Paladin",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 339984,
-    abilityIcon: "item_holyspark.jpg",
-    name: {
-      en: "Focused Light",
-      cn: "聚焦圣光",
-      de: "Fokussiertes Licht",
-      fr: "Lumière concentrée",
-      ru: "Сфокусированный свет",
-    },
-    type: "Potency",
-    class: "Holy Paladin",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 339712,
-    abilityIcon: "ability_priest_voidshift.jpg",
-    name: {
-      en: "Resplendent Light",
-      cn: "壮丽圣光",
-      de: "Prächtiges Licht",
-      fr: "Lumière resplendissante",
-      ru: "Сияющий Свет",
-    },
-    type: "Potency",
-    class: "Holy Paladin",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 339987,
-    abilityIcon: "achievement_admiral_of_the_light.jpg",
-    name: {
-      en: "Untempered Dedication",
-      cn: "痴迷殉道",
-      de: "Ungezügelte Hingabe",
-      fr: "Dévouement intact",
-      ru: "Неумеренная преданность",
-    },
-    type: "Potency",
-    class: "Holy Paladin",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 340218,
-    abilityIcon: "spell_animabastion_orb.jpg",
-    name: {
-      en: "Ringing Clarity",
-      cn: "明澈之音",
-      de: "Läutende Klarheit",
-      fr: "Clarté retentissante",
-      ru: "Звенящая ясность",
-    },
-    type: "Potency",
-    class: "Holy Paladin",
-    covenant: "Kyrian",
-    show: true,
-  },
-  {
-    guid: 340212,
-    abilityIcon: "spell_animarevendreth_groundstate.jpg",
-    name: {
-      en: "Hallowed Discernment",
-      cn: "圣灰洞察",
-      de: "Geheiligte Einsicht",
-      fr: "Discernement sacré",
-      ru: "Священная проницательность",
-    },
-    type: "Potency",
-    class: "Holy Paladin",
-    covenant: "Venthyr",
-    show: true,
-  },
-  {
-    guid: 340192,
-    abilityIcon: "spell_animamaldraxxus_missile.jpg",
-    name: {
-      en: "Righteous Might",
-      cn: "正义之力",
-      de: "Rechtschaffene Macht",
-      fr: "Puissance pieuse",
-      ru: "Мощь праведников",
-    },
-    type: "Potency",
-    class: "Holy Paladin",
-    covenant: "Necrolord",
-    show: true,
-  },
-  {
-    guid: 340185,
-    abilityIcon: "spell_animaardenweald_buff.jpg",
-    name: {
-      en: "The Long Summer",
-      cn: "漫漫长夏",
-      de: "Der lange Sommer",
-      fr: "Été prolongé",
-      ru: "Долгое лето",
-    },
-    type: "Potency",
-    class: "Holy Paladin",
-    covenant: "Night Fae",
-    show: true,
-  },
-
-  /* --------------------------- Endurance Conduits --------------------------- */
-
-  {
-    guid: 338741,
-    abilityIcon: "ability_racial_holyresistance.jpg",
-    name: {
-      en: "Divine Call",
-      cn: "神圣召唤",
-      de: "Göttlicher Ruf",
-      fr: "Appel divin",
-      ru: "Божественный зов",
-    },
-    type: "Endurance",
-    class: "Holy Paladin",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 339114,
-    abilityIcon: "ability_priest_cascade.jpg",
-    name: {
-      en: "Golden Path",
-      cn: "辉金之道",
-      de: "Goldener Pfad",
-      fr: "Voie dorée",
-      ru: "Золотой путь",
-    },
-    type: "Endurance",
-    class: "Holy Paladin",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 338787,
-    abilityIcon: "ability_vehicle_sonicshockwave.jpg",
-    name: {
-      en: "Shielding Words",
-      cn: "盾卫圣令",
-      de: "Abschirmende Worte",
-      fr: "Mots protecteurs",
-      ru: "Спасительное торжество",
-    },
-    type: "Endurance",
-    class: "Holy Paladin",
-    covenant: "",
-    show: true,
-  },
-
-  /* ---------------------------- Finesse Conduits ---------------------------- */
-
-  {
-    guid: 339316,
-    abilityIcon: "achievement_dungeon_heroic_gloryoftheraider.jpg",
-    name: {
-      en: "Echoing Blessings",
-      cn: "回响祝福",
-      de: "Widerhallende Segen",
-      fr: "Bénédictions résonnantes",
-      ru: "Эхо благословений",
-    },
-    type: "Finesse",
-    class: "Holy Paladin",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 339268,
-    abilityIcon: "inv_lightforgedmatrixability_lightsjudgment.jpgg",
-    name: {
-      en: "Light's Barding",
-      cn: "圣光骑甲",
-      de: "Geschirr des Lichts",
-      fr: "Caparaçon de la Lumière",
-      ru: "Доспехи скакуна Света",
-    },
-    type: "Finesse",
-    class: "Holy Paladin",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 339124,
-    abilityIcon: "spell_holy_mindsooth.jpg",
-    name: {
-      en: "Pure Concentration",
-      cn: "纯粹专注",
-      de: "Reine Konzentration",
-      fr: "Pure concentration",
-      ru: "Чистое сосредоточение",
-    },
-    type: "Finesse",
-    class: "Holy Paladin",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 339292,
-    abilityIcon: "spell_holy_painsupression.jpg",
-    name: {
-      en: "Wrench Evil",
-      cn: "扭转邪恶",
-      de: "Böses ausrotten",
-      fr: "Renvoi du mal",
-      ru: "Искоренение зла",
-    },
-    type: "Finesse",
-    class: "Holy Paladin",
-    covenant: "",
-    show: true,
-  },
-
-  /* -------------------------------------------------------------------------- */
-  /*                                   Priest                                   */
-  /* -------------------------------------------------------------------------- */
-
-  /* ---------------------------- Potency Conduits ---------------------------- */
-
-  {
-    guid: 337790,
-    abilityIcon: "spell_holy_spiritualguidence.jpg", // This is a blizzard misspelling, not mine. Just by the way.
-    name: {
-      en: "Exaltation",
-      cn: "欣喜",
-      de: "Verherrlichung",
-      fr: "Exaltation",
-      ru: "Экзальтация",
-    },
-    type: "Potency",
-    class: "Discipline Priest",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 337786,
-    abilityIcon: "spell_holy_blessedrecovery.jpg",
-    name: {
-      en: "Pain Transformation",
-      cn: "痛苦转换",
-      de: "Schmerztransformation",
-      fr: "Transformation de la douleur",
-      ru: "Обращение боли",
-    },
-    type: "Potency",
-    class: "Discipline Priest",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 337778,
-    abilityIcon: "ability_priest_spiritoftheredeemer.jpg",
-    name: {
-      en: "Shining Radiance",
-      cn: "闪耀光辉",
-      de: "Glänzendes Strahlen",
-      fr: "Radiance éclatante",
-      ru: "Яркое сияние",
-    },
-    type: "Potency",
-    class: "Discipline Priest",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 337891,
-    abilityIcon: "spell_holy_purify.jpg",
-    name: {
-      en: "Swift Penitence",
-      cn: "极效苦修",
-      de: "Schnelle Buße",
-      fr: "Prompte pénitence",
-      ru: "Быстрое покаяние",
-    },
-    type: "Potency",
-    class: "Discipline Priest",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 337914,
-    abilityIcon: "achievement_bg_returnxflags_def_wsg.jpg",
-    name: {
-      en: "Focused Mending",
-      cn: "专注愈合",
-      de: "Fokussierte Besserung",
-      fr: "Guérison focalisée",
-      ru: "Направленное исцеление",
-    },
-    type: "Potency",
-    class: "Holy Priest",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 338345,
-    abilityIcon: "ability_priest_bindingprayers.jpg",
-    name: {
-      en: "Holy Oration",
-      cn: "圣言演说",
-      de: "Heilige Ansprache",
-      fr: "Oraison sacrée",
-      ru: "Священная молитва",
-    },
-    type: "Potency",
-    class: "Holy Priest",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 337811,
-    abilityIcon: "inv_icon_wing06a.jpg",
-    name: {
-      en: "Lasting Spirit",
-      cn: "延续之魂",
-      de: "Fortwährender Geist",
-      fr: "Esprit tenace",
-      ru: "Терпеливый дух",
-    },
-    type: "Potency",
-    class: "Holy Priest",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 337947,
-    abilityIcon: "spell_holy_holybolt.jpg",
-    name: {
-      en: "Resonant Words",
-      cn: "共鸣圣言",
-      de: "Resonierende Worte",
-      fr: "Mots résonnants",
-      ru: "Резонирующие слова",
-    },
-    type: "Potency",
-    class: "Holy Priest",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 337966,
-    abilityIcon: "spell_animabastion_wave.jpg",
-    name: {
-      en: "Courageous Ascension",
-      cn: "骁勇晋升",
-      de: "Mutiger Aufstieg",
-      fr: "Ascension courageuse",
-      ru: "Отважное перерождение",
-    },
-    type: "Potency",
-    class: "Priest",
-    covenant: "Kyrian",
-    show: true,
-  },
-  {
-    guid: 338315,
-    abilityIcon: "spell_animarevendreth_debuff.jpg",
-    name: {
-      en: "Shattered Perceptions",
-      cn: "破碎认知",
-      de: "Erschütterte Wahrnehmung",
-      fr: "Perceptions brisées",
-      ru: "Расколотое восприятие",
-    },
-    type: "Potency",
-    class: "Priest",
-    covenant: "Venthyr",
-    show: true,
-  },
-  {
-    guid: 337979,
-    abilityIcon: "spell_animamaldraxxus_debuff.jpg",
-    name: {
-      en: "Festering Transfusion",
-      cn: "溃烂灌注",
-      de: "Schwärende Transfusion",
-      fr: "Transfusion purulente",
-      ru: "Гнилостное переливание",
-    },
-    type: "Potency",
-    class: "Priest",
-    covenant: "Necrolord",
-    show: true,
-  },
-  {
-    guid: 338305,
-    abilityIcon: "spell_animaardenweald_buff.jpg",
-    name: {
-      en: "Fae Fermata",
-      cn: "法夜延音",
-      de: "Feenfermate",
-      fr: "Point d’orgue faë",
-      ru: "Арденвельдская фермата",
-    },
-    type: "Potency",
-    class: "Priest",
-    covenant: "Night Fae",
-    show: true,
-  },
-  {
-    guid: 337715,
-    abilityIcon: "ability_priest_clarityofwill.jpg",
-    name: {
-      en: "Charitable Soul",
-      cn: "慷慨之魂",
-      de: "Großzügige Seele",
-      fr: "Âme charitable",
-      ru: "Щедрая душа",
-    },
-    type: "Endurance",
-    class: "Priest",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 337748,
-    abilityIcon: "spell_holy_restoration.jpg",
-    name: {
-      en: "Light's Inspiration",
-      cn: "圣光鼓舞",
-      de: "Inspiration des Lichts",
-      fr: "Inspiration de la Lumière",
-      ru: "Вдохновение Света",
-    },
-    type: "Endurance",
-    class: "Priest",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 337662,
-    abilityIcon: "spell_nature_invisibilty.jpg",
-    name: {
-      en: "Translucent Image",
-      cn: "通透影像",
-      de: "Durchscheinendes Abbild",
-      fr: "Image translucide",
-      ru: "Прозрачный образ",
-    },
-    type: "Endurance",
-    class: "Priest",
-    covenant: "",
-    show: true,
-  },
-
-  /* ---------------------------- Finesse Conduits ---------------------------- */
-
-  {
-    guid: 337707,
-    abilityIcon: "ability_pvp_innerrenewal.jpg",
-    name: {
-      en: "Clear Mind",
-      cn: "清醒心智",
-      de: "Klarer Kopf",
-      fr: "Esprit clair",
-      ru: "Ясный разум",
-    },
-    type: "Finesse",
-    class: "Priest",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 337954,
-    abilityIcon: "sha_ability_rogue_bloodyeye_nightborne.jpg",
-    name: {
-      en: "Mental Recovery",
-      cn: "精神复建",
-      de: "Mentale Erholung",
-      fr: "Récupération mentale",
-      ru: "Ментальное восстановление",
-    },
-    type: "Finesse",
-    class: "Priest",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 337678,
-    abilityIcon: "ability_priest_savinggrace.jpg",
-    name: {
-      en: "Move with Grace",
-      cn: "优雅步态",
-      de: "Anmutige Bewegung",
-      fr: "Démarche distinguée",
-      ru: "Небесная грация",
-    },
-    type: "Finesse",
-    class: "Priest",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 337762,
-    abilityIcon: "ability_priest_wordsofmeaning.jpg",
-    name: {
-      en: "Power Unto Others",
-      cn: "赋能灌注",
-      de: "Macht den Anderen",
-      fr: "Transfert de puissance",
-      ru: "Сила единства",
-    },
-    type: "Finesse",
-    class: "Priest",
-    covenant: "",
-    show: true,
-  },
-
-  /* -------------------------------------------------------------------------- */
-  /*                                   Shaman                                   */
-  /* -------------------------------------------------------------------------- */
-
-  /* ---------------------------- Potency Conduits ---------------------------- */
-
-  {
-    guid: 338329,
-    abilityIcon: "spell_shaman_improvedearthshield.jpg",
-    name: {
-      en: "Embrace of Earth",
-      cn: "大地之拥",
-      de: "Umarmung der Erde",
-      fr: "Étreinte de la terre",
-      ru: "Объятия Земли",
-    },
-    type: "Potency",
-    class: "Restoration Shaman",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 338343,
-    abilityIcon: "shaman_pvp_ripplingwaters.jpg",
-    name: {
-      en: "Heavy Rainfall",
-      cn: "瓢泼大雨",
-      de: "Schwerer Regen",
-      fr: "Averse torrentielle",
-      ru: "Могучий ливень",
-    },
-    type: "Potency",
-    class: "Restoration Shaman",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 338346,
-    abilityIcon: "spell_shaman_blessingofeternals.jpg",
-    name: {
-      en: "Nature's Reach",
-      cn: "自然专注",
-      de: "Fokus der Natur",
-      fr: "Focalisation de la nature",
-      ru: "Средоточие природы",
-    },
-    type: "Potency",
-    class: "Restoration Shaman",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 338339,
-    abilityIcon: "spell_holy_serendipity.jpg",
-    name: {
-      en: "Swirling Currents",
-      cn: "旋涡之流",
-      de: "Wirbelnde Strömungen",
-      fr: "Courants tourbillonnants",
-      ru: "Кружащие потоки",
-    },
-    type: "Potency",
-    class: "Restoration Shaman",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 339182,
-    abilityIcon: "ability_monk_forcesphere.jpg",
-    name: {
-      en: "Elysian Dirge",
-      cn: "极乐挽歌",
-      de: "Elysisches Klagelied",
-      fr: "Requiem élyséen",
-      ru: "Элизийская панихида",
-    },
-    type: "Potency",
-    class: "Restoration Shaman",
-    covenant: "Kyrian",
-    show: true,
-  },
-  {
-    guid: 339185,
-    abilityIcon: "warlock_charredremains.jpg",
-    name: {
-      en: "Lavish Harvest",
-      cn: "华丽收割",
-      de: "Üppige Ernte",
-      fr: "Moisson généreuse",
-      ru: "Щедрая жатва",
-    },
-    type: "Potency",
-    class: "Restoration Shaman",
-    covenant: "Venthyr",
-    show: true,
-  },
-  {
-    guid: 339186,
-    abilityIcon: "spell_animamaldraxxus_wave.jpg",
-    name: {
-      en: "Tumbling Waves",
-      cn: "翻腾之潮",
-      de: "Stürzende Wellen",
-      fr: "Fracas des vagues",
-      ru: "Волна за волной",
-    },
-    type: "Potency",
-    class: "Restoration Shaman",
-    covenant: "Necrolord",
-    show: true,
-  },
-  {
-    guid: 339183,
-    abilityIcon: "spell_animaardenweald_debuff.jpg",
-    name: {
-      en: "Essential Extraction",
-      cn: "精粹提取",
-      de: "Essenzielle Extraktion",
-      fr: "Extraction essentielle",
-      ru: "Извлечение сути",
-    },
-    type: "Potency",
-    class: "Restoration Shaman",
-    covenant: "Night Fae",
-    show: true,
-  },
-
-  /* --------------------------- Endurance Conduits --------------------------- */
-
-  {
-    guid: 337964,
-    abilityIcon: "ability_hunter_mastertactitian.jpg",
-    name: {
-      en: "Astral Protection",
-      cn: "星界守护",
-      de: "Astralschutz",
-      fr: "Protection astrale",
-      ru: "Астральная защита",
-    },
-    type: "Endurance",
-    class: "Restoration Shaman",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 337974,
-    abilityIcon: "ability_shaman_fortifyingwaters.jpg",
-    name: {
-      en: "Refreshing Waters",
-      cn: "复苏水域",
-      de: "Erfrischende Wasser",
-      fr: "Eaux rafraîchissantes",
-      ru: "Освежающие воды",
-    },
-    type: "Endurance",
-    class: "Restoration Shaman",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 337981,
-    abilityIcon: "ability_accretion.jpg",
-    name: {
-      en: "Vital Accretion",
-      cn: "活力冲积",
-      de: "Vitalzuwachs",
-      fr: "Accrétion vitale",
-      ru: "Прирост жизненных сил",
-    },
-    type: "Endurance",
-    class: "Restoration Shaman",
-    covenant: "",
-    show: true,
-  },
-
-  /* ---------------------------- Finesse Conduits ---------------------------- */
-
-  {
-    guid: 338054,
-    abilityIcon: "spell_nature_removedisease.jpg",
-    name: {
-      en: "Crippling Hex",
-      cn: "致残妖术",
-      de: "Verkrüppelnde Verhexung",
-      fr: "Maléfice affaiblissant",
-      ru: "Увечащий сглаз",
-    },
-    type: "Finesse",
-    class: "Restoration Shaman",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 338048,
-    abilityIcon: "spell_arcane_prismaticcloak.jpg",
-    name: {
-      en: "Spiritual Resonance",
-      cn: "灵魂共鸣",
-      de: "Spirituelle Resonanz",
-      fr: "Résonance spirituelle",
-      ru: "Духовный резонанс",
-    },
-    type: "Finesse",
-    class: "Restoration Shaman",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 338033,
-    abilityIcon: "ability_hunter_longevity.jpg",
-    name: {
-      en: "Thunderous Paws",
-      cn: "雷霆之爪",
-      de: "Donnernde Pfoten",
-      fr: "Pattes tonitruantes",
-      ru: "Громоподобные лапы",
-    },
-    type: "Finesse",
-    class: "Restoration Shaman",
-    covenant: "",
-    show: true,
-  },
-  {
-    guid: 338042,
-    abilityIcon: "spell_shaman_dropall_03.jpg",
-    name: {
-      en: "Totemic Surge",
-      cn: "图腾潮涌",
-      de: "Kraft der Totems",
-      fr: "Afflux totémique",
-      ru: "Сила тотемов",
-    },
-    type: "Finesse",
-    class: "Restoration Shaman",
-    covenant: "",
-    show: true,
-  },
-];
-
-export const soulbindAbilities = [
-  // Kyrian
-  // Pelagos
-  {
-    guid: 328257,
-    icon: "ability_hunter_readiness.jpg",
-    name: "Let Go of the Past",
-  },
-  // Druids
-  {
-    guid: 329777,
-    abilityIcon: "inv_gizmo_khoriumpowercore.jpg",
-    name: "Phial of Patience",
-  },
-  {
-    guid: 329786,
-    abilityIcon: "achievement_guildperk_havegroup-willtravel.jpg",
-    name: "Road of Trials",
-  },
-  {
-    guid: 328261,
-    abilityIcon: "spell_arcane_focusedpower.jpg",
-    name: "Focusing Mantra",
-  },
-  {
-    guid: 328265,
-    abilityIcon: "ability_hunter_displacement.jpg",
-    name: "Bond of Friendship",
-  },
-  {
-    guid: 328263,
-    abilityIcon: "inv_misc_bag_14.jpg",
-    name: "Cleansed Vestments",
-  },
-  {
-    guid: 328266,
-    abilityIcon: "spell_animabastion_buff.jpg",
-    name: "Combat Meditation",
-  },
-
-  // Kleia
-  {
-    guid: 329776,
-    abilityIcon: "inv_alchemy_elixir_02.jpg",
-    name: "Ascendant Phial",
-  },
-  {
-    guid: 334066,
-    abilityIcon: "achievement_reputation_05.jpg",
-    name: "Mentorship",
-  },
-  {
-    guid: 329781,
-    abilityIcon: "spell_holy_mindvision.jpg",
-    name: "Resonant Accolades",
-  },
-  {
-    guid: 329779,
-    abilityIcon: "inv_boots_plate_05.jpg",
-    name: "Bearer's Pursuit",
-  },
-  {
-    guid: 329778,
-    abilityIcon: "ability_hunter_displacement.jpg",
-    name: "Pointed Courage",
-  },
-  {
-    guid: 329784,
-    abilityIcon: "spell_magic_magearmor.jpg",
-    name: "Cleansing Rites",
-  },
-  {
-    guid: 328258,
-    abilityIcon: "spell_frost_wisp.jpg",
-    name: "Ever Forward",
-  },
-  {
-    guid: 329791,
-    abilityIcon: "spell_holy_prayerofspirit.jpg",
-    name: "Valiant Strikes",
-  },
-
-  // Forgelite Prime Mikanikos
-
-  {
-    guid: 331610,
-    abilityIcon: "inv_potion_105.jpg",
-    name: "Charged Additive",
-  },
-  {
-    guid: 331609,
-    abilityIcon: "inv_alchemy_lightbloodelixir.jpg",
-    name: "Forgelite Filter",
-  },
-  {
-    guid: 333935,
-    abilityIcon: "spell_paladin_lightshammer.jpg",
-    name: "Hammer of Genesis",
-  },
-  {
-    guid: 331612,
-    abilityIcon: "inv_jewelcrafting_80_cutgem02_blue.jpg",
-    name: "Sparkling Driftglobe Core",
-  },
-  {
-    guid: 331611,
-    abilityIcon: "inv_misc_enggizmos_05.jpg",
-    name: "Soulsteel Clamps",
-  },
-  {
-    guid: 331726,
-    abilityIcon: "inv_misc_enchantedpearlf.jpg",
-    name: "Regenerating Materials",
-  },
-  {
-    guid: 331725,
-    abilityIcon: "achievement_legionpvp4tier1.jpg",
-    name: "Resilient Plumage",
-  },
-  {
-    guid: 333950,
-    abilityIcon: "ui_sigil_kyrian.jpg",
-    name: "Bron's Call to Action",
-  },
-
-  // Venthyr
-  // Nadjia the Mistblade
-  {
-    guid: 331576,
-    abilityIcon: "spell_shadow_deathsembrace.jpg",
-    name: "Agent of Chaos",
-  },
-  {
-    guid: 331577,
-    abilityIcon: "ability_rogue_sprint_blue.jpg",
-    name: "Fancy Footwork",
-  },
-  {
-    guid: 331579,
-    abilityIcon: "ability_rogue_dirtydeeds.jpg",
-    name: "Friends in Low Places",
-  },
-  {
-    guid: 331580,
-    abilityIcon: "trade_brewpoison.jpg",
-    name: "Exacting Preparation",
-  },
-  {
-    guid: 331582,
-    abilityIcon: "sha_ability_rogue_bloodyeye_nightmare.jpg",
-    name: "Familiar Predicaments",
-  },
-  {
-    guid: 331584,
-    abilityIcon: "inv_tabard_duelersguild.jpg",
-    name: "Dauntless Duelist",
-  },
-  {
-    guid: 331586,
-    abilityIcon: "ability_rogue_bloodsplatter.jpg",
-    name: "Thrill Seeker",
-  },
-
-  // Theotar the Mad Duke
-
-  {
-    guid: 336140,
-    abilityIcon: "inv_raidpriestmythic_q_01boot.jpg",
-    name: "Watch the Shoes!",
-  },
-  {
-    guid: 336147,
-    abilityIcon: "spell_lifegivingspeed.jpg",
-    name: "Leisurely Gait",
-  },
-  {
-    guid: 336243,
-    abilityIcon: "inv_trinket_revendreth_01_gold.jpg",
-    name: "Refined Palate",
-  },
-  {
-    guid: 336239,
-    abilityIcon: "archaeology_5_0_umbrellaofchiji.jpg",
-    name: "Soothing Shade",
-  },
-  {
-    guid: 336245,
-    abilityIcon: "inv_misc_coin_01.jpg",
-    name: "Token of Appreciation",
-  },
-  {
-    guid: 336247,
-    abilityIcon: "achievement_halloween_smiley_01.jpg",
-    name: "Life of the Party",
-  },
-  {
-    guid: 336150,
-    abilityIcon: "trade_archaeology_queenazshara-dressinggown.jpg",
-    name: "Impeccable Style",
-  },
-  {
-    guid: 336184,
-    abilityIcon: "achievement_guildperk_bountifulbags.jpg",
-    name: "Exquisite Ingredients",
-  },
-  {
-    guid: 319983,
-    abilityIcon: "inv_drink_23.jpg",
-    name: "Wasteland Propriety",
-  },
-
-  // General Draven
-
-  {
-    guid: 332755,
-    abilityIcon: "inv_chest_mail_revendreth_d_01.jpg",
-    name: "Unbreakable Body",
-  },
-  {
-    guid: 332862,
-    abilityIcon: "spell_misc_warsongfocus.jpg",
-    name: "Expedition Leader",
-  },
-  {
-    guid: 319973,
-    abilityIcon: "ability_warrior_intensifyrage.jpg",
-    name: "Built for War",
-  },
-  {
-    guid: 332753,
-    abilityIcon: "inv_inscription_80_warscroll_battleshout.jpg",
-    name: "Superior Tactics",
-  },
-  {
-    guid: 332754,
-    abilityIcon: "achievement_bg_killxenemies_generalsroom.jpg",
-    name: "Hold Your Ground",
-  },
-  {
-    guid: 319978,
-    abilityIcon: "ability_garrosh_touch_of_yshaarj.jpg",
-    name: "Enduring Gloom",
-  },
-  {
-    guid: 319982,
-    abilityIcon: "spell_holy_prayerofhealing.jpg",
-    name: "Move As One",
-  },
-  {
-    guid: 340159,
-    abilityIcon: "inv_misc_reforgedarchstone_01.jpg",
-    name: "Service In Stone",
-  },
-
-  // Necrolord
-
-  //Plague Deviser Marileth
-
-  {
-    guid: 323074,
-    abilityIcon: "inv_misc_food_legion_gooslime_pool.jpg",
-    name: "Volatile Solvent",
-  },
-  {
-    guid: 323089,
-    abilityIcon: "inv_maldraxxusslime_red.jpg",
-    name: "Travel with Bloop",
-  },
-  {
-    guid: 323091,
-    abilityIcon: "inv_maldraxxusslime_green.jpg",
-    name: "Ooz's Frictionless Coating",
-  },
-  {
-    guid: 323090,
-    abilityIcon: "inv_maldraxxusslime_yellow.jpg",
-    name: "Plaguey's Preemptive Strike",
-  },
-  {
-    guid: 323079,
-    abilityIcon: "inv_maldraxxusslime_purple.jpg",
-    name: "Kevin's Keyring",
-  },
-  {
-    guid: 323081,
-    abilityIcon: "inv_maldraxxusslime_plague.jpg",
-    name: "Plagueborn Cleansing Slime",
-  },
-  {
-    guid: 323095,
-    abilityIcon: "spell_shadow_zdemonformgreen.jpg",
-    name: "Ultimate Form",
-  },
-
-  // Emeni, the Slaughter Daughter
-
-  {
-    guid: 323921,
-    abilityIcon: "inv_helm_crown_c_01_black.jpg",
-    name: "Emeni's Magnificent Skin",
-  },
-  {
-    guid: 341650,
-    abilityIcon: "inv_helm_crown_c_01_rose.jpg",
-    name: "Emeni's Ambulatory Flesh",
-  },
-  {
-    guid: 324440,
-    abilityIcon: "inv_pants_leather_raiddruid_j_01.jpg",
-    name: "Cartilaginous Legs",
-  },
-  {
-    guid: 324441,
-    abilityIcon: "item_alchemiststonec.jpg",
-    name: "Hearth Kidneystone",
-  },
-  {
-    guid: 323918,
-    abilityIcon: "spell_nzinsanity_leadenfoot.jpg",
-    name: "Gristled Toes",
-  },
-  {
-    guid: 323919,
-    abilityIcon: "ability_creature_poison_01.jpg",
-    name: "Gnashing Chompers",
-  },
-  {
-    guid: 323916,
-    abilityIcon: "spell_shadow_plaguecloud.jpg",
-    name: "Sulfuric Emission",
-  },
-  {
-    guid: 342156,
-    abilityIcon: "trade_archaeology_vrykul_runestick.jpg",
-    name: "Lead by Example",
-  },
-
-  // Bonesmith Heirmir
-
-  {
-    guid: 326504,
-    abilityIcon: "inv_misc_sawblade_01.jpg",
-    name: "Serrated Spaulders",
-  },
-  {
-    guid: 326507,
-    abilityIcon: "inv_cooking_80_brinyflesh.jpg",
-    name: "Resourceful Fleshcrafting",
-  },
-  {
-    guid: 326509,
-    abilityIcon: "spell_deathknight_empowerruneblade.jpg",
-    name: "Heirmir's Arsenal: Ravenous Pendant",
-  },
-  {
-    guid: 326511,
-    abilityIcon: "inv_boots_raidshaman_i_01.jpg",
-    name: "Heirmir's Arsenal: Gorestompers",
-  },
-  {
-    guid: 326572,
-    abilityIcon: "inv_misc_enchantedpearla.jpg",
-    name: "Heirmir's Arsenal: Marrowed Gemstone",
-  },
-  {
-    guid: 326512,
-    abilityIcon: "trade_archaeology_bones-of-transformation.jpg",
-    name: "Runeforged Spurs",
-  },
-  {
-    guid: 326513,
-    abilityIcon: "inv_misc_coinbag14.jpg",
-    name: "Bonesmith's Satchel",
-  },
-  {
-    guid: 326514,
-    abilityIcon: "garrison_greenarmor.jpg",
-    name: "Forgeborne Reveries",
-  },
-
-  // Night Fae
-  // Niya
-
-  {
-    guid: 342270,
-    abilityIcon: "ability_racial_runningwild.jpg",
-    name: "Run Without Tiring",
-  },
-  {
-    guid: 320658,
-    abilityIcon: "inv_misc_foot_centaur.jpg",
-    name: "Stay on the Move",
-  },
-  {
-    guid: 320659,
-    abilityIcon: "ability_rogue_trip.jpg",
-    name: "Niya's Tools: Burrs",
-  },
-  {
-    guid: 320660,
-    abilityIcon: "trade_brewpoison.jpg",
-    name: "Niya's Tools: Poison",
-  },
-  {
-    guid: 320662,
-    abilityIcon: "inv_misc_herb_arthastear_petal.jpg",
-    name: "Niya's Tools: Herbs",
-  },
-  {
-    guid: 320668,
-    abilityIcon: "inv_misc_bag_27.jpg",
-    name: "Nature's Splendor",
-  },
-  {
-    guid: 320687,
-    abilityIcon: "ability_druid_supriseattack.jpg",
-    name: "Swift Patrol",
-  },
-  {
-    guid: 322721,
-    abilityIcon: "ability_druid_manatree.jpg",
-    name: "Grove Invigoration",
-  },
-  // Dreamweaver
-
-  {
-    guid: 319191,
-    abilityIcon: "inv_misc_trailofflowers.jpg",
-    name: "Field of Blossoms",
-  },
-  {
-    guid: 319210,
-    abilityIcon: "spell_shaman_blessingofeternals.jpg",
-    name: "Social Butterfly",
-  },
-  {
-    guid: 319211,
-    abilityIcon: "spell_nature_sleep.jpg",
-    name: "Soothing Voice",
-  },
-  {
-    guid: 319213,
-    abilityIcon: "ability_druid_dreamstate.jpg",
-    name: "Empowered Chrysalis",
-  },
-  {
-    guid: 319214,
-    abilityIcon: "inv_misc_dust.jpg",
-    name: "Faerie Dust",
-  },
-  {
-    guid: 319216,
-    abilityIcon: "ability_xavius_dreamsimulacrum.jpg",
-    name: "Somnambulist",
-  },
-  {
-    guid: 319217,
-    abilityIcon: "inv_farm_herbseed.jpg",
-    name: "Podtender",
-  },
-
-  // Korayn
-  {
-    guid: 325067,
-    abilityIcon: "inv_misc_monsterhorn_05.jpg",
-    name: "Horn of the Wild Hunt",
-  },
-  {
-    guid: 325065,
-    abilityIcon: "ability_druid_dash.jpg",
-    name: "Wild Hunt's Charge",
-  },
-  {
-    guid: 325068,
-    abilityIcon: "ability_warrior_endlessrage.jpg",
-    name: "Face Your Foes",
-  },
-  {
-    guid: 325069,
-    abilityIcon: "ability_rogue_surpriseattack.jpg",
-    name: "First Strike",
-  },
-  {
-    guid: 325601,
-    abilityIcon: "spell_nature_shamanrage.jpg",
-    name: "Hold the Line",
-  },
-  {
-    guid: 325072,
-    abilityIcon: "inv_polearm_2h_heirloomspear_c_01.jpg",
-    name: "Vorkai Sharpening Techniques",
-  },
-  {
-    guid: 325073,
-    abilityIcon: "inv_misc_foot_centaur.jpg",
-    name: "Get In Formation",
-  },
-  {
-    guid: 325066,
-    abilityIcon: "ability_blackhand_marked4death.jpg",
-    name: "Wild Hunt Tactics",
-  },
-];
 
 export const classRaceList = {
   "Holy Paladin": {
