+++ src/Modules/Covenants/Components/CovenantExploration.js
@@ -2,7 +2,7 @@ import React, { useEffect } from "react";
 import PropTypes from "prop-types";
 import { makeStyles } from "@material-ui/core/styles";
 import { Tabs, Tab, Box, AppBar, Grid, Paper, Typography } from "@material-ui/core";
-import { soulbindDB, soulbindConnectors } from "../SoulbindDB";
+import { soulbindDB, soulbindConnectors } from "../../../Databases/SoulbindDB";
 import SoulbindNode from "./SoulbindNode";
 import ConduitObject from "./ConduitObject";
 import { getSoulbindFormula } from "../../Engine/EffectFormulas/Generic/GenericSoulbindFormulas";
