+++ src/Modules/UpgradeFinder/UpgradeFinderEngine.js
@@ -1,4 +1,4 @@
-import { itemDB, tokenDB } from "../Player/ItemDB";
+import { itemDB, tokenDB } from "../../Databases/ItemDB";
 import Item from "../Player/Item";
 import { runTopGear } from "../TopGear/TopGearEngine";
 import {
